% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_snapped_points_strahler.R
\name{api_get_snapped_points_strahler}
\alias{api_get_snapped_points_strahler}
\title{Snap Points to River Network Using Strahler Threshold}
\usage{
api_get_snapped_points_strahler(
  df,
  colname_lat = "latitude",
  colname_lon = "longitude",
  colname_site_id = "site_id",
  strahler = 3,
  geometry_only = FALSE,
  comment = NULL,
  process_url = NULL
)
}
\arguments{
\item{df}{A data.frame containing point data.}

\item{colname_lat}{Name of the latitude column. Default: "latitude".}

\item{colname_lon}{Name of the longitude column. Default: "longitude".}

\item{colname_site_id}{Name of the site ID column. Default: "site_id".}

\item{strahler}{Minimum Strahler order to snap to (e.g., 3).}

\item{geometry_only}{Logical. If TRUE, return only geometry.}

\item{comment}{Optional comment string passed to the API.}

\item{process_url}{URL of the pygeoapi process (e.g., "http://localhost:5000/processes/get-snapped-points-strahler/execution").}
}
\value{
A data.frame with the snapped point data.
}
\description{
Sends individual coordinate points to the GeoFRESH API to snap them
to the nearest stream segment above a given Strahler order.
}
\seealso{
Other ocgapi: 
\code{\link{api_get_env90m_data}()},
\code{\link{api_get_local_ids}()},
\code{\link{api_get_snapped_points}()},
\code{\link{api_get_upstream_subcids}()}
}
\concept{ocgapi}
