% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/utils_download.R
\name{check_wsl}
\alias{check_wsl}
\title{Check if WSL and Ubuntu is installed on Windows}
\usage{
check_wsl()

check_wsl()
}
\description{
Check if WSL and Ubuntu is installed on Windows

Check if WSL and Ubuntu is installed on Windows
}
\keyword{internal}
