% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph_utils.R
\name{choose_hydro_timeline}
\alias{choose_hydro_timeline}
\title{Choose Timeline for Hydrograph (Calendar Year or Water Year)}
\usage{
choose_hydro_timeline(
  all_hydro_sites_historical,
  all_hydro_sites_1yr,
  WY,
  locations,
  location_num
)
}
\arguments{
\item{all_hydro_sites_historical}{Data frame of historical hydrometric data.}

\item{all_hydro_sites_1yr}{Data frame of current year hydrometric data.}

\item{WY}{Character string indicating whether to use calendar year ("no") or water year ("yes").}

\item{locations}{WSC locations}

\item{location_num}{number of iteration through loop}
}
\value{
A list containing the filtered historical and current year datasets, and the year label.
}
\description{
This function selects and filters hydrometric data based on whether the data should
be plotted by calendar year or water year.
}
