% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_wrangle.R
\name{clean_climate}
\alias{clean_climate}
\title{Downloads and wrangles hydro data}
\usage{
clean_climate(climate_raw)
}
\arguments{
\item{climate_raw}{A data frame containing raw data downloaded from the
Canadian climate extraction tool}
}
\value{
A clean data frame that can be used in \code{\link[=precip_stats_daily]{precip_stats_daily()}}, \code{\link[=temperature_stats_daily]{temperature_stats_daily()}},
and \code{\link[=temp_plotting_historical]{temp_plotting_historical()}}
}
\description{
\code{clean_climate} This wrangles \emph{daily} climate data acquired from the Canadian
climate extraction tool
(https://climate-change.canada.ca/climate-data/#/hourly-climate-data) and gets
it ready for analysis.
}
\examples{
# Read raw climate csv files and bind into one data frame
# Files should be "daily" climate data from climate extraction tool 
files <- get_example_file()
file_paths <- get_example_file(grep("merritt_", files, value = TRUE))
data_list <- lapply(file_paths, read.csv)
climate_raw <- do.call(rbind, data_list)
#clean_climate(climate_raw)

}
