% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_raw}
\alias{climate_raw}
\title{Raw daily climate data for station MERRITT STP}
\format{
\subsection{\code{who}}{

A data frame with 18,945 rows and 36 columns:
\describe{
\item{x, y}{latitude and longitude in decimal degrees}
\item{SPEED_MAX_GUST_FLAG}{etc etc}
\item{MIN_TEMPERATURE}{etc etc}
...
}
}
}
\source{
\url{https://climate-change.canada.ca/climate-data/#/hourly-climate-data}
}
\usage{
climate_raw
}
\description{
Example \emph{daily} climate data acquired from the Canadian climate extraction tool
(https://climate-change.canada.ca/climate-data/#/hourly-climate-data)
}
\details{
This data frame can be used in \code{\link[=clean_climate]{clean_climate()}}.
}
\keyword{datasets}
