% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_clean_and_prepare_data.R
\name{combine_csv_tidyhydat}
\alias{combine_csv_tidyhydat}
\title{Create a Comprehensive Dataframe of Locations and Date Ranges}
\usage{
combine_csv_tidyhydat(tidyhydat_hydro, csv_Q_dl)
}
\arguments{
\item{tidyhydat_hydro}{A dataframe containing hydrological data retrieved
using the \code{tidyhydat} package.}

\item{csv_Q_dl}{A dataframe containing discharge data manually downloaded
from the WSC website.}
}
\value{
A single dataframe that combines both sources, including updated information
on the relevant date ranges and start/end dates for the data.
}
\description{
\code{combine_csv_tidyhydat} merges discharge data downloaded from the WSC website
with data obtained through the \code{tidyhydat} package, creating a single comprehensive
dataframe. This function is particularly useful for ensuring that all relevant
data, including manually downloaded datasets, are consolidated into one dataset
for further analysis.
}
\details{
This function ensures no data gaps by integrating manually downloaded datasets
with those retrieved programmatically from \code{tidyhydat}. Any discrepancies in
date ranges between the two sources are resolved to provide a complete and
accurate dataset.
}
\examples{
# Example 1: Download and combine data

# Step 1: Retrieve data using tidyhydat
tidy_data <- dl_hydro(station_number = "08HB034", nick_name = "Nanaimo River")

# Step 2: Prepare WSC data
file_path <- get_example_file("Nanaimo_08HB034_QR_20240104T2154.csv")
wsc_data <- reformat_Q_dl_daily(file_path, "Nanaimo River")

# Step 3: Combine datasets
combined_data <- combine_csv_tidyhydat(tidy_data, wsc_data)

}
