% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_results_sf}
\alias{combine_results_sf}
\title{Combine lists of geometries and data into an sf object}
\usage{
combine_results_sf(geoms_list, data_list, crs)
}
\arguments{
\item{geoms_list}{List of LINESTRING geometries}

\item{data_list}{List of data frames corresponding to each geometry}

\item{crs}{Coordinate reference system for output sf}
}
\value{
sf object combining all geometries and data
}
\description{
Safely combines a list of geometries and corresponding attribute data frames
into a single sf object. Ensures there is only one geometry column.
}
\keyword{internal}
