% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_create_stats.R
\name{create_hydro_stats_historical}
\alias{create_hydro_stats_historical}
\alias{create_hydro_stats_singleYr}
\alias{create_hydro_stats_from_Q_dl}
\title{Create Hydrology Summary Statistics}
\usage{
create_hydro_stats_historical(hydro_data, date_minimum = NA, date_maximum = NA)

create_hydro_stats_singleYr(hydro_data, YOI, WY = FALSE)

create_hydro_stats_from_Q_dl(Q_csv_path)
}
\arguments{
\item{hydro_data}{A dataframe containing hydrology data in the format output by \code{dl_hydro}.}

\item{date_minimum}{A character string specifying the earliest date for the summary,
in "yyyy-mm-dd" format. Defaults to the earliest date in the dataset.}

\item{date_maximum}{A character string specifying the latest date for the summary,
in "yyyy-mm-dd" format. Defaults to the latest date in the dataset.}

\item{YOI}{An integer specifying a single year for statistical summarization.
This is used to create the YOI trace for the hydrograph.}

\item{WY}{A logical value indicating whether to filter data by water year
(\code{TRUE}) or calendar year (\code{FALSE}).}

\item{Q_csv_path}{A character string specifying the file path to a CSV containing
manually downloaded discharge data from the WSC. The dataset should represent a
single year of data for hydrograph generation.}
}
\value{
A dataframe containing hydrological statistics summarized by station and day.
}
\description{
The following functions summarize hydrological statistics:
\itemize{
\item \strong{\code{create_hydro_stats_historical}}: Summarizes historical hydrological statistics
into a dataframe, grouped by month and day.
\item \strong{\code{create_hydro_stats_singleYr}}: Generates statistical summaries for a single
year of data. This function filters a specific year from the historical data and
produces corresponding statistics.
\item \strong{\code{create_hydro_stats_from_Q_dl}}: Processes manually downloaded discharge data
(from the WSC website) to generate statistics for a single year. This may be
useful if recent data is not yet available in \code{tidyhydat}.
}

\strong{Note}: All data from February 29 has been excluded from calculations to address
complications related to leap years. This avoids statistical biases caused by only
a subset (1/4) of years including February 29 in the dataset.
}
\examples{
# Example 1: Historical statistics
hydro_data <- dl_hydro(station_number = "08LD001", nickname = "Adams River")

# Generate historical statistics
create_hydro_stats_historical(hydro_data)

# Specify date range for summary
create_hydro_stats_historical(hydro_data, date_minimum = "2015-11-02", 
                               date_maximum = "2021-12-31")

# Example 2: Single year statistics
create_hydro_stats_singleYr(hydro_data, YOI = 2021, WY = FALSE)

}
