% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_download.R
\name{dl_hydro}
\alias{dl_hydro}
\alias{reformat_Q_dl_daily}
\title{Download and Reformat Hydrology Data}
\usage{
dl_hydro(station_number, nick_name = NULL)

reformat_Q_dl_daily(Q_csv_path, nick_name = NULL)
}
\arguments{
\item{station_number}{A single station number or a list of station numbers to download data for.}

\item{nick_name}{A single nickname or a list of nicknames corresponding to \code{station_number}.
Must match the length and order of \code{station_number}. Use \code{NA} if no nicknames are required.}

\item{Q_csv_path}{A character string specifying the path to a CSV file containing daily
hydrology data manually downloaded from the WSC website.}
}
\value{
A dataframe containing hydrologic data for all specified stations in long format.
}
\description{
\itemize{
\item \strong{\code{dl_hydro}}: Downloads hydrology data for specified station(s) using the \code{tidyhydat}
package, optionally appends a nickname to each station, and outputs the data in long format.
\item \strong{\code{reformat_Q_dl_daily}}: Reformats manually downloaded daily hydrology data
(from the WSC website) into the same format as data downloaded through \code{tidyhydat},
ensuring consistency for analysis.
}
}
\examples{
# Example 1: Download data for a single station
dl_hydro(station_number = "08LD001", nick_name = "Adams River")

# Example 2: Download data for multiple stations
dl_hydro(station_number = c("08LD001", "08LC002"), 
         nick_name = c("Adams River", "Shuswap River"))

# Example 3: Reformat manually downloaded data
file_path <- get_example_file("Nanaimo_08HB034_QR_20240104T2154.csv")
reformat_Q_dl_daily(file_path)
}
