% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_env90m_tables.R
\name{download-env90m-tables}
\alias{download-env90m-tables}
\alias{download_soil_tables}
\alias{download_flo1k_tables}
\alias{download_cgiar_tables}
\alias{download_merit_dem_tables}
\alias{download_hydrography90m_tables}
\alias{download_present_climate_tables}
\alias{download_future_climate_tables}
\alias{download_landcover_tables}
\title{Download Environment90m tables}
\usage{
download_soil_tables(
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = NULL
)

download_flo1k_tables(
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = NULL
)

download_cgiar_tables(
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = NULL
)

download_merit_dem_tables(
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = NULL
)

download_hydrography90m_tables(
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = NULL
)

download_present_climate_tables(
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = NULL
)

download_future_climate_tables(
  base_vars = NULL,
  time_periods = NULL,
  models = NULL,
  scenarios = NULL,
  versions = NULL,
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = NULL
)

download_landcover_tables(
  base_vars = NULL,
  years = NULL,
  subset = NULL,
  tile_ids = NULL,
  download = FALSE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE,
  ignore_missing = FALSE,
  tempdir = NULL,
  quiet = FALSE
)
}
\arguments{
\item{subset}{Vector of the variable names that should be downloaded (or
string "ALL" for all available variables).}

\item{tile_ids}{Vector containing all tile ids of the tiles (e.g. "h10v04")
that should be downloaded, or whose download availability and size should
be checked(or string "ALL" for all available tiles).}

\item{download}{logical. If TRUE, and if `tile_ids` is specified, the files
will be downloaded from the IGB server. If FALSE, and if `tile_ids` is
specified, the download size will be computed. If FALSE, and if `tile_ids`
is NULL, the variable names will be returned to the user.}

\item{download_dir}{Directory where the downloads should be stored.
Defaults to the current working directory ".". Ignored if `download=FALSE`.}

\item{file_format}{File format of the tables, either "txt" or "zip". If
"txt", then the zipped tables are unzipped. If "zip", the downloaded
zipped files are left as they are. Default is "txt", which means that the
zip files are unzipped after downloading. Note that this will take more
space on disk than zips.}

\item{delete_zips}{logical. boolean If `FALSE`, the downloaded zip files are
not deleted after unzipping. Defaults to TRUE. This is ignored if you
request file format zip.}

\item{ignore_missing}{logical. What to do if some of the requested variables
and/or tile_ids are not available, which is most frequently caused by a
typo the variable name. If TRUE, the missing or misspelled ones are
ignored while the others are downloaded (and a warning is given out).
If FALSE, the function will fail to allow the user to check the variable
names and their spelling. Defaults to FALSE.}

\item{tempdir}{Optional (rarely needed). Path to the directory where to
store/look for the temporary various file size tables for the various
Environment90m datasets, which are required and downloaded by the
functions. If not passed, defaults to the output of [base::tempdir()].}

\item{quiet}{logical. If FALSE, informative messages will be printed.
Default is FALSE.}

\item{base_vars}{(Only in `download_future_climate_tables()` and
`download_landcover_tables()`) Vector of the desired base variables, e.g.
the landcover variable "c20_1992" can be expressed as base variable "c20"
and year "1992".}

\item{time_periods}{(Only in `download_future_climate_...`) Vector of the
desired time periods (leave `NULL` or specify `"ALL"` for all available
time periods).}

\item{models}{(Only in `download_future_climate_...`) Vector of the desired
models (leave `NULL` or specify `"ALL"` for all available models).}

\item{scenarios}{(Only in `download_future_climate_...`) Vector of the
desired scenarios (leave `NULL` or specify `"ALL"` for all available
scenarios).}

\item{versions}{(Only in `download_future_climate_...`) Vector of the
desired versions (leave `NULL` or specify `"ALL"` for all available
versions). As of January 2025, the only available version is "V.2.1".}

\item{years}{(Only in `download_landcover_...`) Vector of the desired years
(leave `NULL` or specify `"ALL"` for all available years).}
}
\value{
A named list of:
* variable names,
* components of variable names (if applies),
* the dataset name,
* the requested tile_ids (if applies),
* the download size (if `tile_id` is specified),
* the path where downloaded files are stored to (if `download=TRUE`),
* etc.
}
\description{
The various `download_something_tables()` functions allow to retrieve
the Environment90m variable names and download data of the
Environment90m datasets, which are split into 20°x20° tiles.

If the functions are called without arguments (i.e. without specifying
variable names and tiles), the available variable names are returned.

If a subset of variables and tile IDs are specified, the download size of
the resulting download will be computed.

If a subset of variables and tile IDs are specified, and `download` is set
to `TRUE`, the requested tables will be downloaded and either left as
zipped files or unzipped to text files.

Multiple regular tiles, e.g. belonging to regional units, can be
downloaded in a single request. The tile IDs can be obtained using the
function [get_tile_id()].
}
\details{
In the following table you can find all the variables included in the
Environment90m dataset. The column "Variable" includes the variable names
that should be used as an input in the parameter "variable" of the function.
Likewise, the column "File format" contains the input that should be given to
the "file_format" parameter.

The Environment90m dataset comprises data from the landcover dataset
ESA Land Cover (esa_cci_landcover_v2_1_1), CHELSA v2.1 (chelsa_bioclim_v2_1),
SOILGRIDS (soilgrids250m_v2_0) and Hydrography90m (hydrography90m_v1_0).

For visualisations the available tiles, and for details on the variables of
the Hydrography90m dataset, please refer to
\url{https://hydrography.org/hydrography90m/hydrography90m_layers/}.

For details on the bioclimatic variables, especially for details of the scale
and unit of the values, please refer to
\url{http://chelsa-climate.org/}.

For details on the ESA Land Cover variables, please refer to
\url{https://www.climatologylab.org/terraclimate.html}. Please note that some
values in this dataset are aggregated from similar classes (see 
Environment90m publication).

For details on the Soil data, please refer to
\url{https://soilgrids.org}.
}
\section{Functions}{
\itemize{
\item \code{download_soil_tables()}: Download tables for the 
soilgrids250m_v2_0 dataset

\item \code{download_flo1k_tables()}: Download tables for the
flo1k_v1_0 dataset

\item \code{download_cgiar_tables()}: Download tables for the
cgiar_csi_v3 dataset.

\item \code{download_merit_dem_tables()}: Download tables for the
merit_dem_v1_0_3 dataset.

\item \code{download_hydrography90m_tables()}: Download tables for the 
hydrography90m_v1_0 dataset.

\item \code{download_present_climate_tables()}: Download tables for the
chelsa_bioclim_v2_1 dataset (bioclimatic variables, except for future
simulations)

\item \code{download_future_climate_tables()}: Download tables for the
chelsa_bioclim_v2_1 dataset (bioclimatic variables, future simulations only)

\item \code{download_landcover_tables()}: Download tables for the 
esa_cci_landcover_v2_1_1 dataset

}}
\examples{

### Soil: soilgrids250m_v2_0 ###
# Show all available soil variable names:
download_soil_tables()

# Compute download size of all soil variables, for one tile:
download_soil_tables(
  subset = "ALL",
  tile_ids = c("h00v04"),
  download = FALSE)

# Download one soil variable (Clay content), for two tiles:
download_soil_tables(
  subset = c("clyppt"),
  tile_ids = c("h00v04", "h10v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "zip")

# Download one soil variable (Clay content), for one tile,
# unzip, and delete the zips:
download_soil_tables(
  subset = c("clyppt"),
  tile_ids = c("h00v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE) 


### Flow (mean): flo1k_v1_0 ###
# Show all available flo1k variable names:
download_flo1k_tables()

# Compute download size of the only flo1k_v1_0 variable (mean flow),
# for one tile:
download_flo1k_tables(
  subset = "ALL",
  tile_ids = c("h00v04"),
  download = FALSE)

# Download the only flo1k_v1_0 variable (flo1k), for two tiles:
download_soil_tables(
  subset = c("flo1k"),
  tile_ids = c("h00v04", "h10v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "zip")

# Download the only flo1k_v1_0 variable (flo1k), for one tile:
# unzip, and delete the zips:
download_soil_tables(
  subset = c("flo1k"),
  tile_ids = c("h00v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE)


### CGIAR CSI dataset: cgiar_csi_v3 ###
# Show all available cgiar variable names
download_cgiar_tables()

# Compute download size of all cgiar variables, for one tile:
download_cgiar_tables(
  subset = "ALL",
  tile_ids = c("h00v04"),
  download = FALSE)

# Download one cgiar variable (Global Aridity Index), for two tiles:
download_cgiar_tables(
  subset = c("garid"),
  tile_ids = c("h00v04", "h10v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "zip")

# Download two cgiar variables (Global Aridity Index, Potential
# Evapotranspiration), for one tile, unzip, and delete the zips:
download_cgiar_tables(
  subset = c("garid", "gevapt"),
  tile_ids = c("h00v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE)


### Digital Elevation Model: merit_dem_v1_0_3 ###
# (Multi-Error-Removed Improved-Terrain Digital Elevation Model)
# Show all available merit_dem_v1_0_3 variable names
download_merit_dem_tables()

# Compute download size of the only merit_dem_v1_0_3 variable
# (Mean elevation), for one tile:
download_merit_dem_tables(
  subset = "ALL",
  tile_ids = c("h00v04"),
  download = FALSE)

# Download the only merit_dem_v1_0_3 variable (mean elevation),
# for two tiles:
download_merit_dem_tables(
  subset = c("elev"), # or "ALL"
  tile_ids = c("h00v04", "h10v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "zip")

# Download the only merit_dem_v1_0_3 variable (mean elevation),
# for one tile, unzip, and delete the zips:
download_merit_dem_tables(
  subset = c("elev"),
  tile_ids = c("h00v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE)


### Hydrography90m: hydrography90m_v1_0 ###
# Show all available hy90m variable names
download_hydrography90m_tables()

# Compute download size of all hy90m variables, for one tile:
download_hydrography90m_tables(
  subset = "ALL",
  tile_ids = c("h00v04"),
  download = FALSE)

# Download one hy90m variable (Strahler’s stream order), for two tiles:
download_hydrography90m_tables(
  subset = c("stream_strahler"),
  tile_ids = c("h00v04", "h10v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "zip")

# Download one hy90m variable (Strahler’s stream order), for one tile,
# unzip, and delete the zips:
download_hydrography90m_tables(
  subset = c("stream_strahler"),
  tile_ids = c("h00v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE) 


### Bioclimatic Variables: chelsa_bioclim_v2_1 ###
### (excluding future simulations)             ###
# Show all available bioclim variable names
# (excluding future simulations):
download_present_climate_tables()

# Compute download size of all bioclim variables, for one tile:
download_present_climate_tables(
  subset = "ALL",
  tile_ids = c("h00v04"),
  download = FALSE)

# Download one bioclim variable (Annual mean temperature), for two tiles:
download_present_climate_tables(
  subset = c("bio1"),
  tile_ids = c("h00v04", "h10v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "zip")

# Download one bioclim variable (Annual mean temperature), for one tile,
# unzip, and delete the zips:
download_present_climate_tables(
  subset = c("bio1"),
  tile_ids = c("h00v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE) 


### Bioclimatic Variables: chelsa_bioclim_v2_1 ###
### (future simulations)                       ###    
# Show all available future bioclim variable names
download_future_climate_tables()

# Compute download size of all variables, for one tile:
download_future_climate_tables(
  subset = "ALL",
  tile_ids = c("h00v04"),
  download = FALSE)

# Download one hy90m variable (Annual mean temperature), for two tiles:
download_future_climate_tables(
  subset = c("bio1"),
  tile_ids = c("h00v04", "h10v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "zip")

# Download one hy90m variable (Annual mean temperature), for one tile,
# unzip, and delete the zips:
download_future_climate_tables(
  subset = c("bio1"),
  tile_ids = c("h00v04"),
  download = TRUE,
  download_dir = ".",
  file_format = "txt",
  delete_zips = TRUE) 


### Landcover: esa_cci_landcover_v2_1_1 ###
# Show all available landcover variable names:
download_landcover_tables()

# Compute download size of two landcover base variables (Cropland, rainfed,
# and Grassland) and two years, for all tiles:
  vars <- download_landcover_tables(
    base_vars=c("c10", "c130"),
    years=c(1992, 1993),
    tile_ids="ALL")

# Download two base variables and one year, for two tiles:
  vars <- download_landcover_tables(
    base_vars=c("c10", "c130"),
    years=c(1992),
    tile_ids=c("h00v04", "h10v04"),
    download=TRUE,
    download_dir="/tmp",
    file_format="zip",
    delete_zips=FALSE)

}
\references{
Garcia Marquez J., Amatulli, G., Grigoropoulou, A.,
Schürz, M., Tomiczek, T., Buurman, M., Bremerich, V.,
Bego, K. and Domisch, S.:
Global datasets of aggregated environmental variables at the
sub-catchment scale for freshwater biodiversity modeling, in prep.
Please contact the authors for more up-to-date citation info.
}
\seealso{
[download_tiles()] for downloading spatial layers (raster, vector)
  of the original Hydrography90m dataset, split to the same tiles.
}
\author{
Merret Buurman
}
