% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_env.R
\name{download_env}
\alias{download_env}
\title{Download files of the Environment90m dataset}
\usage{
download_env(
  variable,
  file_format = "txt",
  years = NULL,
  tile_id = NULL,
  download_dir = ".",
  delete_zips = TRUE
)
}
\arguments{
\item{variable}{character vector of variable names. See Details for all the
variable names.}

\item{file_format}{character. Format of the requested file. "zip" and "txt"
are supported. Default is "txt", which means that the zip files are unzipped
after downloading. Note that this will take more space on disk than zips.}

\item{years}{character vector of years. Only valid for ESA Land Cover variables,
ignored otherwise.}

\item{tile_id}{character vector. The IDs of the requested tiles.}

\item{download_dir}{character. The directory where the files will be
downloaded. Default is the working directory.}

\item{delete_zips}{boolean If FALSE, the downloaded zip files are not deleted
after unzipping. Defaults to TRUE. This is ignored if you request file format zip.}
}
\description{
The function downloads data of the Environment90m
dataset, which is split into 20°x20° tiles. If a tile ID is specified, then
the selected layers (variable) will be downloaded.
Multiple regular tiles, e.g. belonging to regional units, can be
downloaded in a single request. The tile or regional unit IDs can be
obtained using the functions "get_tile_id" and "get_regional_unit_id",
respectively.
The files will be stored locally in a folder architecture, similar as in the
data repository, available at
\url{https://public.igb-berlin.de/index.php/s/agciopgzXjWswF4?path=\%2F}.
}
\details{
In the following table you can find all the variables included in the
Environment90m dataset. The column "Variable" includes the variable names
that should be used as an input in the parameter "variable" of the function.
Likewise, the column "File format" contains the input that should be given to
the "file_format" parameter.

For more details and visualisations of the spatial layers (e.g. the available
tiles), please refer to
\url{https://hydrography.org/hydrography90m/hydrography90m_layers/}.
For details on the bioclimatic variables, especially for details of the scale
and unit of the values, please refer to
\url{http://chelsa-climate.org/}.
For details on the ESA Land Cover variables, please refer to
\url{https://www.climatologylab.org/terraclimate.html}. Please note that some
values in this dataset are aggregated from similar classes (see table below,
see Environment90m publication).
For details on the Soil data, please refer to
\url{https://soilgrids.org}.

TODO: The CHELSA units also have scale and offset! Explain them here?
TODO: The ESA LandCover data are aggregated. Document!\tabular{lllll}{
   \strong{Variable type} \tab \strong{Variable name} \tab \strong{Variable} \tab \strong{Unit} \tab \strong{File format} \cr
   Bioclimatic variables (present) \tab Annual Mean Temperature \tab bio1 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Mean Diurnal Range \tab bio2 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Isothermality \tab bio3 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Temperature Seasonality \tab bio4 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Max Temperature of Warmest Month \tab bio5 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Min Temperature of Coldest Month \tab bio6 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Temperature Annual Range \tab bio7 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Mean Temperature of Wettest Quarter \tab bio8 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Mean Temperature of Driest Quarter \tab bio9 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Mean Temperature of Warmest Quarter \tab bio10 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Mean Temperature of Coldest Quarter \tab bio11 \tab degrees Celsius \tab zip / txt \cr
   Bioclimatic variables (present) \tab Annual Precipitation \tab bio12 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (present) \tab Precipitation of Wettest Month \tab bio13 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (present) \tab Precipitation of Driest Month \tab bio14 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (present) \tab Precipitation Seasonality \tab bio15 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (present) \tab Precipitation of Wettest Quarter \tab bio16 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (present) \tab Precipitation of Driest Quarter \tab bio17 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (present) \tab Precipitation of Warmest Quarter \tab bio18 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (present) \tab Precipitation of Coldest Quarter \tab bio19 \tab kg/m^2 \tab zip / txt \cr
   Bioclimatic variables (2071-2100) \tab Same as above \tab  \tab  \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Cropland, rainfed (aggregated from original classes 10,11,12) \tab c10_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Cropland, irrigated or post-flooding \tab c20_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Mosaic cropland (>50\%) / natural vegetation (tree, shrub, herbaceous cover) (<50\%) \tab c30_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Mosaic natural vegetation (tree, shrub, herbaceous cover) (>50\%) / cropland (<50\%) \tab c40_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Tree cover, broadleaved, evergreen, closed to open (>15\%) \tab c50_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Tree cover, broadleaved, deciduous, closed to open (>15\%) (aggregated from original classes 60, 61, 62) \tab c60_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Tree cover, needleleaved, evergreen, closed to open (>15\%) (aggregated from original classes 70, 71, 72) \tab c70_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Tree cover, needleleaved, deciduous, closed to open (>15\%) (aggregated from original classes 80, 81, 82) \tab c80_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Tree cover, mixed leaf type (broadleaved and needleleaved) \tab c90_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Mosaic tree and shrub (>50\%) / herbaceous cover (<50\%) \tab c100_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Mosaic herbaceous cover (>50\%) / tree and shrub (<50\%) \tab c110_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Shrubland (aggregated from original classes 120, 121, 122) \tab c120_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Grassland \tab c130_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Lichens and mosses \tab c140_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Sparse vegetation (tree, shrub, herbaceous cover) (<15\%) (aggregated from original classes 150, 151, 152, 153) \tab c150_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Tree cover, flooded, fresh or brackish water \tab c160_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Tree cover, flooded, saline water \tab c170_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Shrub or herbaceous cover, flooded, fresh/saline/brackish water \tab c180_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Urban areas (aggregated from original classes 200, 201, 202) \tab c190_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Bare areas \tab c200_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Water bodies \tab c210_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   ESA Land Cover (1992-2018) \tab Permanent snow and ice \tab c220_\if{html}{\out{<year>}} \tab \% \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Depth to bedrock (R horizon) up to 200cm \tab soil_BDRICM \tab cm \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Grade of a sub-soil being acid \tab soil_ACDWRB \tab pH \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Derived saturated water content \tab soil_AWCTS \tab  \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Clay content \tab soil_CLYPPT \tab \% \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Sand content \tab soil_SNDPPT \tab \% \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Silt content \tab soil_SLTPPT \tab \% \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Derived available soil water capacity \tab soil_WWP \tab  \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Texture class (USDA system) \tab soil_TEXMHT \tab  \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Soil organic carbon content (fine earth fraction) \tab soil_ORCDRC \tab g/kg \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Soil pH x 10 in H2O \tab soil_PHIHOX \tab pH \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Bulk density (fine earth) \tab soil_BLDFIE \tab kg / cubic-meter \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Cation exchange capacity \tab soil_CECSOL \tab cmolc/kg \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Coarse fragments volumetric \tab soil_CRFVOL \tab \% \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Grade of a sub-soil being acid e.g. having a pH < 5 and low BS \tab soil_ACDWRB \tab ph \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Probability of occurence of R horizon (0-100\%) \tab soil_BDRLOG \tab \% \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Cummulative probability of organic soil based on the TAXOUSDA and TAXNWRB \tab soil_HISTPR \tab  \tab zip / txt \cr
   Soil (Global gridded soil information) \tab Sodic soil grade based on WRB soil types and soil pH \tab soil_SLGWRB \tab pH \tab zip / txt \cr
}
}
\note{
If there is an error during the download of a file
(more likely in case of files bigger than 3-4GB), you can try to manually
download this file by pasting the link that is returned by the error
message in your browser.
}
\examples{
# Define directory
my_directory <- tempdir()
# Download data for two variables in three regular tiles
# to the current working directory
download_env(variable = c("bio1", "bio2"),
               tile_id = c("h00v02", "h16v02", "h16v04"),
               download_dir = my_directory)

# Download them as zip, to save disk space:
download_env(variable = c("bio1", "bio2"),
               file_format = "zip",
               tile_id = c("h00v02", "h16v02", "h16v04"),
               download_dir = my_directory)

# Download land cover data for two years
download_env(variable = c("c20", "c30"), years = c(1992, 1996),
               file_format = "zip",
               tile_id = c("h00v02","h16v02", "h16v04"),
               download_dir = my_directory)
# Download land cover data for specific variables and  years
download_env(variable = c("c20_1996", "c30_1992"),
               file_format = "zip",
               tile_id = c("h00v02", "h16v02", "h16v04"),
               download_dir = my_directory)
}
\references{
Amatulli G., Garcia Marquez J., Sethi T., Kiesel J.,
Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 )
Hydrography90m: A new high-resolution global hydrographic dataset.
IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries.
dataset. \url{https://doi.org/10.18728/igb-fred-762.1}
}
\author{
Afroditi Grigoropoulou, Merret Buurman
}
