% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_future_climate.R
\name{download_future_climate}
\alias{download_future_climate}
\title{Download files of the Environment90m dataset}
\usage{
download_future_climate(
  variable,
  file_format = "csv",
  time_period = NULL,
  scenario = NULL,
  model = NULL,
  tile_id = NULL,
  version = "V.2.1",
  download_dir = ".",
  delete_zips = TRUE
)
}
\arguments{
\item{variable}{character vector of variable names. See Details for all the
variable names. No default, has to be provided.}

\item{file_format}{character. Format of the requested file. "zip" and "csv"
are supported. Default is "csv", which means that the zip files are unzipped
after downloading. Note that this will take more space on disk than zips.}

\item{time_period}{character vector of time periods. Default: "2071-2100".
Currently only "2071-2100" is available, so this can be omitted.}

\item{scenario}{character vector of scenarios, e.g. "ssp370", "ssp585". See
Details for all the scenario names. No default, has to be provided.}

\item{model}{character vector of models, e.g. "mpi-esm1-2-hr", "ukesm1-0-ll",
"ipsl-cm6a-lr". See Details for all the scenario names. No default, has to be
provided.}

\item{tile_id}{character vector. The IDs of the requested tiles.}

\item{version}{string. e.g. "V.2.1". Defaults to "V.2.1".}

\item{download_dir}{character. The directory where the files will be
downloaded. Default is the working directory.}

\item{delete_zips}{boolean If FALSE, the downloaded zip files are not deleted
after unzipping. Defaults to TRUE. This is ignored if you request file format zip.}
}
\description{
The function downloads data of the Environment90m
dataset, which is split into 20°x20° tiles. If a tile ID is specified, then
the selected layers (variable) will be downloaded.
Multiple regular tiles, e.g. belonging to regional units, can be
downloaded in a single request. The tile or regional unit IDs can be
obtained using the functions "get_tile_id" and "get_regional_unit_id",
respectively.
The files will be stored locally in a folder architecture, similar as in the
data repository, available at
\url{https://public.igb-berlin.de/index.php/s/agciopgzXjWswF4?path=\%2F}.
}
\details{
In the following table you can find all the variables included in the
Environment90m dataset. The column "Variable" includes the variable names
that should be used as an input in the parameter "variable" of the function.
Likewise, the column "File format" contains the input that should be given to
the "file_format" parameter.
For more details and visualisations of the spatial layers (e.g. the available
tiles), please refer to
\url{https://hydrography.org/hydrography90m/hydrography90m_layers/}.
For details on the bioclimatic variables, especially for details of the scale
and unit of the values, please refer to
\url{http://chelsa-climate.org/}.

TODO: The CHELSA units also have scale and offset! Explain them here?\tabular{lllll}{
   \strong{Variable type} \tab \strong{Variable name} \tab \strong{Variable} \tab \strong{Unit} \tab \strong{File format} \cr
   Bioclimatic variables (2071-2100) \tab Annual Mean Temperature \tab bio1 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Mean Diurnal Range \tab bio2 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Isothermality \tab bio3 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Temperature Seasonality \tab bio4 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Max Temperature of Warmest Month \tab bio5 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Min Temperature of Coldest Month \tab bio6 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Temperature Annual Range \tab bio7 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Mean Temperature of Wettest Quarter \tab bio8 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Mean Temperature of Driest Quarter \tab bio9 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Mean Temperature of Warmest Quarter \tab bio10 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Mean Temperature of Coldest Quarter \tab bio11 \tab degrees Celsius \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Annual Precipitation \tab bio12 \tab kg/m^2 \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Precipitation of Wettest Month \tab bio13 \tab kg/m^2 \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Precipitation of Driest Month \tab bio14 \tab kg/m^2 \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Precipitation Seasonality \tab bio15 \tab kg/m^2 \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Precipitation of Wettest Quarter \tab bio16 \tab kg/m^2 \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Precipitation of Driest Quarter \tab bio17 \tab kg/m^2 \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Precipitation of Warmest Quarter \tab bio18 \tab kg/m^2 \tab zip / csv \cr
   Bioclimatic variables (2071-2100) \tab Precipitation of Coldest Quarter \tab bio19 \tab kg/m^2 \tab zip / csv \cr
}


| \strong{Scenario name} | \strong{Scenario description}   |
| ssp370            | TODO: Scenario description |
| ssp585            | TODO: Scenario description |

| \strong{Model name} | \strong{Scenario description} |
| mpi-esm1-2-hr  | TODO: Model description  |
| psl-cm6a-lr    | TODO: Model description  |
| ukesm1-0-ll    | TODO: MOdel description  |

TODO: Also add something on versions!
}
\note{
If there is an error during the download of a file
(more likely in case of files bigger than 3-4GB), you can try to manually
download this file by pasting the link that is returned by the error
message in your browser.
}
\examples{
# Download bioclimatic variables for specific variables, scenarios, models:
variable = c("bio1")
scenario = c("ssp370")
model = c("ipsl-cm6a-lr")
time_period = c("2071-2100")
download_future_climate(variable = variable, file_format = "zip",
               scenario = scenario, model = model,
               time_period = time_period,
               tile_id = c("h00v02", "h16v02", "h16v04"))

# Download many bioclimatic variables:
variable = c("bio1", "bio2", "bio3", "bio4", "bio5", "bio6", "bio7", "bio8",
             "bio9", "bio10", "bio11", "bio12", "bio13", "bio14", "bio15",
             "bio16", "bio17", "bio18", "bio19")
scenario = c("ssp370", "ssp585")
model = c("ipsl-cm6a-lr", "mpi-esm1-2-hr", "ukesm1-0-ll")
time_period = c("2071-2100")
download_future_climate(variable = variable, file_format = "zip",
               scenario = scenario, model = model,
               time_period = time_period,
               tile_id = c("h00v02", "h16v02", "h16v04"))

}
\references{
Amatulli G., Garcia Marquez J., Sethi T., Kiesel J.,
Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 )
Hydrography90m: A new high-resolution global hydrographic dataset.
IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries.
dataset. \url{https://doi.org/10.18728/igb-fred-762.1}
}
\author{
Afroditi Grigoropoulou, Merret Buurman
}
