% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tiles.R
\name{download_tiles}
\alias{download_tiles}
\title{Download files of the Hydrography90m dataset}
\usage{
download_tiles(
  variable,
  file_format = "tif",
  tile_id = NULL,
  reg_unit_id = NULL,
  global = FALSE,
  download_dir = "."
)
}
\arguments{
\item{variable}{character vector of variable names.}

\item{file_format}{character. Format of the requested file ("tif" or "gpkg").}

\item{tile_id}{character vector. The IDs of the requested tiles.}

\item{reg_unit_id}{character vector. The IDs of the requested regional units.}

\item{global}{logical. If TRUE, the global extent file is downloaded.
Default is FALSE.}

\item{download_dir}{character. The directory where the files will be
downloaded. Default is the working directory.}
}
\description{
The function downloads files of the Hydrography90m
dataset, available at https://public.igb-berlin.de/index.php/s/agciopgzXjWswF4?path=\%2F .
The files will be stored in the folder architecture of the above domain.
Multiple regular tile or regional unit files can be requested in a single
call of the function. The tile or regional unit IDs can be obtained
using the functions "get_tile_id" and "get_regional_unit_id" respectively.
}
\details{
\tabular{lllll}{
   \strong{Variable type} \tab \strong{Variable} \tab \strong{Abbreviation} \tab \strong{Unit} \tab \strong{Description} \cr
   Flow \tab Flow accumulation \tab flow \tab km² \tab Flow accumulation \cr
   Flow \tab Flow accumulation (positive values) \tab flowpos \tab km² \tab Flow accumulation with positive values \cr
   Stream slope \tab Cell maximum curvature \tab slope_curv_max_dw_cel \tab m⁻¹ \tab Cell maximum curvature (between highest upstream cell,focal cell and downstream cell) \cr
   Stream slope \tab Cell minimum curvature \tab slope_curv_min_dw_cel \tab m⁻¹ \tab Cell minimum curvature(between lowest upstream cell,focal cell and downstream cell) \cr
   Stream slope \tab Cell elevation difference \tab slope_elv_dw_cel \tab m \tab Cell elevation difference(between focal cell and downstream cell) \cr
   Stream slope \tab Cell gradient \tab slope_grad_dw_cel \tab  \tab Cell gradient \cr
   Stream distance \tab Shortest distance to drainage divide \tab stream_dist_up_near \tab m \tab Shortest upstream distance between focal grid cell and the nearest sub-catchment drainage divide \cr
   Stream distance \tab Longest distance to drainage divide \tab stream_dist_up_farth \tab m \tab Longest upstream distance between focal grid cell and the nearest sub-catchment drainage divide \cr
   Stream distance \tab Nearest down stream stream grid cell \tab stream_dist_dw_near \tab m \tab Distance between focal grid cell and its nearest down stream stream grid cell \cr
   Stream distance \tab Outlet grid cell in the network \tab outlet_dist_dw_basin \tab m \tab Distance between focal grid cell and the outlet grid cell in the network \cr
   Stream distance \tab Down stream stream node grid cell \tab outlet_dist_dw_scatch \tab m \tab Distance between focal grid cell and the down stream stream node grid cell \cr
   Stream distance \tab Euclidean distance \tab stream_dist_proximity \tab m \tab Euclidean distance between focal grid cell and the stream network (in meters) \cr
   Elevation difference \tab Shortest path \tab stream_diff_up_near \tab m \tab Elevation difference of the shortest path from focal grid cell to the sub-catchment drainage divide \cr
   Elevation difference \tab Longest path \tab stream_diff_up_farth \tab m \tab Elevation difference of the longest path from focal grid cell to the sub-catchment drainage divide \cr
   Elevation difference \tab Nearest downstream stream pixel \tab stream_diff_dw_near \tab m \tab Elevation difference between focal grid cell and its nearest downstream stream pixel \cr
   Elevation difference \tab Outlet grid cell in the network \tab outlet_diff_dw_basin \tab m \tab Elevation difference between focal grid cell and the outlet grid cell in the network \cr
   Elevation difference \tab Downstream stream node grid cell \tab outlet_diff_dw_scatch \tab m \tab Elevation difference between focal grid cell and the downstream stream node grid cell \cr
   Segment properties \tab Segment downstream mean gradient \tab channel_grad_dw_seg \tab  \tab Segment downstream mean gradient (between focal cell and the node/outlet) \cr
   Segment properties \tab Segment upstream mean gradient \tab channel_grad_up_seg \tab  \tab Segment upstream mean gradient (between focal cell and the init/node) \cr
   Segment properties \tab Cell upstream gradient \tab channel_grad_up_cel \tab  \tab Cell upstream gradient (between focal cell and next cell \cr
}


\if{html}{\out{<div class="sourceCode">}}\preformatted{                                                                                                                               | Segment properties   | Cell stream course curvature             | channel curv_cel      |          | Cell stream course curvature (focal cell)                                                           |
                                                                                                                               | Segment properties   | Segment downstream elevation difference  | channel_elv_dw_seg    |          | Segment downstream elevation difference (between focal cell and the node/outlet)                    |
                                                                                                                               | Segment properties   | Segment upstream elevation difference    | channel_elv_up_seg    |          | Segment upstream elevation difference (between focal cell and the init/node)                        |
                                                                                                                               | Segment properties   | Cell upstream elevation difference       | channel_elv_up_cel    |          | Cell upstream elevation difference (between focal cell and next cell)                               |
                                                                                                                               | Segment properties   | Cell downstream elevation difference     | channel_elv_dw_cel    |          | Cell downstream elevation difference (between focal cell and next cell)                             |
                                                                                                                               | Segment properties   | Segment downstream distance              | channel_dist_dw_seg   |          | Segment downstream distance (between focal cell and the node/outlet                                 |
                                                                                                                                                                                                                                                                     | Segment properties   | Segment upstream distance                | channel_dist_up_seg   |          | Segment upstream distance (between focal cell and the init/node)                                    |
                                                                                                                                                                                                                                                                     | Segment properties   | Cell upstream distance                   | channel_dist_up_cel   |          | Cell upstream distance (between focal cell and next cell)                                           |
                                                                                                                                                                                                                                                                     | Stream order         | Strahler’s stream order                  | stream_strahler       |          | Strahler’s stream order                                                                             |
                                                                                                                                                                                                                                                                     | Stream order         | Shreve’s stream magnitude                | stream_shreve         |          | Shreve’s stream magnitude                                                                           |
                                                                                                                                                                                                                                                                     | Stream order         | Horton’s stream order                    | stream_horton         |          | Horton’s stream order                                                                               |
                                                                                                                                                                                                                                                                     | Stream order         | Hack’s stream order                      | stream_hack           |          | Hack’s stream order                                                                                 |
                                                                                                                                                                                                                                                                     | Stream order         | Topological dimension of streams         | stream_topo           |          | Topological dimension of streams                                                                    |
                                                                                                                                                                                                                                                                     | Stream reach         | Length of the stream reach               | length                |          | Length of the stream reach                                                                          |
                                                                                                                                                                                                                                                                     | Stream reach         | Straight length                          | stright               |          | Length of stream as straight line                                                                   |
                                                                                                                                                                                                                                                                     | Stream reach         | Sinusoid of the stream reach             | sinusoid              |          | Fractal dimension: stream length/straight stream length                                             |
                                                                                                                                                                                                                                                                     | Stream reach         | Accumulated length                       | cum_length            |          | Length of stream from source                                                                        |
                                                                                                                                                                                                                                                                     | Stream reach         | Flow accumulation (?)                    | flow_accum            |          | Flow accumulation (what is the difference to the one above???)                                      |
                                                                                                                                                                                                                                                                     | Stream reach         | Distance to outlet                       | out_dist              |          | Distance of current stream init from outlet                                                         |
                                                                                                                                                                                                                                                                     | Stream reach         | Source elevation                         | source_elev           |          | Elevation of stream init                                                                            |
                                                                                                                                                                                                                                                                     | Stream reach         | Outlet elevation                         | outlet_elev           |          | Elevation of stream outlet                                                                          |
                                                                                                                                                                                                                                                                     | Stream reach         | Elevation drop                           | elev_drop             |          | Difference between source_elev and outlet_elev + drop outlet                                        |
                                                                                                                                                                                                                                                                     | Stream reach         | Outlet drop                              | out_drop              |          | Drop at the outlet of the stream                                                                    |
                                                                                                                                                                                                                                                                     | Stream reach         | Gradient                                 | gradient              |          | Mean gradient of the sub-catchment (downstream elevation difference divided by distance)            |
                                                                                                                                                                                                                                                                     | Flow index           | Stream power index                       | spi                   |          | A measure of the erosive power of flowing water (Moore et al., 1991)                                |
                                                                                                                                                                                                                                                                     | Flow index           | Sediment transportation index            | sti                   |          | A metric for describing the erosion and deposition of sediments (Mojaddadi et al., 2017)            |
                                                                                                                                                                                                                                                                     | Flow index           | Compound topographic index               | cti                   |          | "A steady state wetness index, also known as topographic wetness index (TWI) (Beven & Kirkby,       |
}\if{html}{\out{</div>}}

| 1979)"               |                                          |                       |          |                                                                                                     |
}
\examples{
# Download data for two variables in three regular tiles
# to the current working directory
download_tiles(variable = c("sti", "stream_dist_up_farth"),
               file_format = "tif",
               tile_id = c("h00v02","h16v02", "h16v04"))

# Download the global .tif layer for the variable "direction"
# into the temporary R folder or define a different directory
# Define directory
my_directory <- tempdir()
# Download layer
download_tiles(variable = "direction",
               file_format = "tif",
               global = TRUE,
               download_dir = my_directory)

}
\references{
Amatulli G., Garcia Marquez J., Sethi T., Kiesel J.,
Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 )
Hydrography90m: A new high-resolution global hydrographic dataset.
IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries.
dataset. https://doi.org/10.18728/igb-fred-762.1
}
\author{
Afroditi Grigoropoulou
}
