% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ids.R
\name{extract_ids}
\alias{extract_ids}
\title{Adds sub-catchment and/or basin IDs to a dataframe of points}
\usage{
extract_ids(
  data,
  lon = character(0),
  lat = character(0),
  subcatchment_path = NULL,
  basin_path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{data}{Data.frame with lat/lon columns}

\item{lon}{Name of the longitude column (WGS 84)}

\item{lat}{Name of the latitude column (WGS 84)}

\item{subcatchment_path}{Full path to the sub-catchment ID .tif layer}

\item{basin_path}{Full path to the basin ID .tif layer}

\item{quiet}{If FLASE, process is printed}
}
\description{
Adds sub-catchment and/or basin IDs to a dataframe of points
}
