% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_stat.R
\name{extract_var_stat}
\alias{extract_var_stat}
\title{Calculate statistics based on environmental variable raster layers}
\usage{
extract_var_stat(subc_ids, subc_layer, variables, n_cores)
}
\arguments{
\item{dataset_path}{Path to data set text file with the lat long columns}

\item{lon}{Name of the longitude column}

\item{lat}{Name of the latitude column}

\item{subc_path}{Path to the sub-catchment ID .tif layer}

\item{basin_path}{Path to the basin ID .tif layer}
}
\description{
Add the column 'subcID' to the dataset in 'dataset_path' and export the
data frame in a csv.
In the current version the long-lat values must be in the 2nd and 3rd
column of the data set text file.
}
