% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_zonal_stat.R
\name{extract_zonal_stat}
\alias{extract_zonal_stat}
\title{Calculate zonal statistics based on environmental variable raster and vector layers.}
\usage{
extract_zonal_stat(data_dir, out_path = NULL, subc_ids, subc_layer, variables)
}
\arguments{
\item{data_dir}{Path to the directory containing all input data}

\item{out_path}{Full path of the output file. If not NULL, the output data.frame is exported as a csv in the given path}

\item{subc_ids}{Vector of sub-catchment ids, can be acquired from the extract_ids() function, by sub setting the resulting data.frame}

\item{subc_layer}{Full path to the sub-catchment ID .tif layer}

\item{variables}{Variable file names, e.g. slope_grad_dw_cel_h00v00.tif. Variable names should remain intact in file names, even after prior file processing, i.e., slope_grad_dw_cel should appear in the file name}
}
\description{
Calculate zonal statistics based on environmental variable raster and vector layers.
}
