% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_utils.R
\name{get_custom_plot_seq}
\alias{get_custom_plot_seq}
\title{Creates a custom seq for axis for plotting}
\usage{
get_custom_plot_seq(custom_ymax, custom_ymin)
}
\arguments{
\item{custom_ymax}{ymax of data}

\item{custom_ymin}{ymin of data}
}
\value{
A sequence of numbers to be used for x or y axis on a plot
}
\description{
\code{get_custom_plot_seq} uses custom_ymax, custom_ymin inputs to create appropriate
sequence of numbers.
}
