% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ranges.R
\name{get_date_ranges}
\alias{get_date_ranges}
\title{Creates a df with each location and the date range of the data}
\usage{
get_date_ranges(all_hydro_sites_combined)
}
\arguments{
\item{all_hydro_sites_combined}{df of all the hydrology data}
}
\value{
A dataframe with updated information on the date range of each location
}
\description{
\code{get_date_ranges} Uses the location names and the combined hydrologic data to
create and then output a df that contains the date ranges for each location
}
