% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{Calculate euclidean or along the network distance between points}
\usage{
get_distance(
  data,
  lon,
  lat,
  id,
  basin_id = NULL,
  basin_layer = NULL,
  stream_layer = NULL,
  distance = "all",
  n_cores = 1,
  quiet = TRUE
)
}
\arguments{
\item{data}{a data.frame or data.table with lat/lon coordinates in WGS84.}

\item{lon}{character. The name of the column with the longitude coordinates.}

\item{lat}{character. The name of the column with the latitude coordinates.}

\item{id}{character. The name of a column containing unique IDs for each row
of "data" (e.g., occurrence or site IDs).}

\item{basin_id}{character. The name of the column with the basin IDs.
If NULL and distance is set to 'network' or 'both', the basin IDs will be
extracted automatically. Default is NULL.}

\item{basin_layer}{character. Full path to the basin ID .tif layer.}

\item{stream_layer}{character. Full path of the stream network .gpkg file.}

\item{distance}{character. One of "euclidean", "network", or "both".
If "euclidean", the euclidean distances between all pairs of points are
calculated. If "network", the shortest path along the network between all
pairs of points is calculated. (see "Details" for more information).
If method is set to "both", both distance measures are calculated.
Default is "both".}

\item{n_cores}{numeric. Number of cores used for parallelization.
Default is 1.}

\item{quiet}{logical. If FALSE, the standard output will be printed.
Default is TRUE.}
}
\value{
If distance='euclidean', a distance matrix of the euclidean distances between
all the pairs of points (object of class data.frame). If distance='network',
a data.frame with three columns: from_id, to_id, dist. The 'dist' column
includes the distance of the shortest path along the network from the node
from_id to the node to_id. If distance='both', a list containing both objects
is returned.
}
\description{
Calculate euclidean or along the network distance between points
}
\details{
...
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Load occurrence data
species_occurrence <- read.table(paste0(my_directory,
                            "/hydrography90m_test_data/spdata_1264942.txt"),
                              header = TRUE)

basin_rast <- paste0(my_directory,
                     "/hydrography90m_test_data/basin_1264942.tif")

# Define full path to the sub-catchment raster layer
basin_rast <- paste0(my_directory,
                     "/hydrography90m_test_data/subcatchment_1264942.tif")

# Define full path to the vector file of the stream network
stream_vect <- paste0(my_directory,
                      "/hydrography90m_test_data/order_vect_59.gpkg")

# Automatically extract the basin and sub-catchment IDs and
# snap the data points to the stream segment
snapped_coordinates <- snap_to_subc_segment(data = species_occurrence,
                                            lon = "longitude",
                                            lat = "latitude",
                                            id = "occurrence_id",
                                            basin_layer = basin_rast,
                                            subc_layer = subc_rast,
                                            stream_layer = stream_vect,
                                            n_cores = 2)
# Show head of output table
head(snapped_coordinates)

# Get the euclidean distance and the distance along the network between all
# pairs of points
distance_table <- get_distance(data = snapped_coordinates,
                               lon = "lon_snap",
                               lat = "lat_snap",
                               id = "occurrence_id",
                               basin_id = "basin_id",
                               basin_layer = basin_rast,
                               stream_layer = stream_vect,
                               distance = "network",
                               n_cores = 1,
                               quiet = FALSE)
}
\references{
\url{https://grass.osgeo.org/grass82/manuals/v.net.allpairs.html}
\url{https://grass.osgeo.org/grass82/manuals/v.distance.html}
}
\seealso{
\code{\link{snap_to_network}} to snap the data points to the next stream
segment within a given radius and/or a given flow accumulation threshold
value.
\code{\link{extract_ids}} to extract basin and sub-catchment IDs.
}
\author{
Afroditi Grigoropoulou, Maria M. Üblacker, Jaime Garcia Marquez
}
