% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrographr-example-data.R
\name{get_example_file}
\alias{get_example_file}
\title{Access Example Data Files in hydroGraphR}
\usage{
get_example_file(path = NULL)
}
\arguments{
\item{path}{Character string specifying the name of the example file. If \code{NULL},
all example files will be listed. Default is \code{NULL}.}
}
\value{
A character vector containing either the full path to the specified file
or a list of all available example files if \code{path} is \code{NULL}.
}
\description{
This utility function provides easy access to example data files included with
the \code{hydroGraphR} package, located in its \code{inst/extdata} directory.
}
\details{
Adapted from https://readxl.tidyverse.org/reference/readxl_example.html
}
\examples{
get_example_file()
get_example_file("kamloops_aut_1163842_1.csv")

files <- get_example_file()
get_example_file(grep("merritt_", files, value = TRUE))

}
