% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_catchment.R
\name{get_lake_catchment}
\alias{get_lake_catchment}
\title{Get lake catchments based on intersection points between stream network and lake}
\usage{
get_lake_catchment(
  data,
  flow = "flow_accu_mean",
  lake_id = "lake_ID",
  outlet_id = "outlet_ID",
  top = "all",
  direction,
  catch,
  n_cores = 1,
  quiet = TRUE
)
}
\arguments{
\item{data}{a data.frame or data.table that contains the columns regarding
the stream segment ids of the intersection points between lake and stream network;
(i.e., output of get_lake_intersection)}

\item{flow}{character. The name of the flow accumulation column used for sorting the intersection table;
i.e. either flow_accu (flow accumulation value at intersection point pixel),
flow_accu_max (maximum flow accumulation value of 3 x 3 neighboring pixels),}

\item{lake_id}{character. The name of the column containing lake ids;
(i.e., output "lake_ID" of get_lake_intersection)
flow_accu_mean (mean flow accumulation value of 3 x 3 neighboring pixels). Default is flow_accu_mean}

\item{top}{integer. Number of intersection points used for lake catchment calculations;
e.g. top=1 equals first row of intersection table and lake outlet;
top=20 equals first 20 rows of intersection table; Default is all.}

\item{direction}{character. Full path to Hydrography 90m flow direction tif file}

\item{catch}{character. Full path to output catchment tif files}

\item{n_cores}{integer. Number of cores used in parallelsation; Default is one.}

\item{quiet}{logical. If FALSE, the standard output will be printed.;
Default is TRUE.}

\item{read}{logical. If TRUE, then the model .csv table
gets read into R as data.table and data.frame.
if FALSE, the table is only stored on disk. Default is FALSE.}
}
\description{
Add describtion text here

 Any thing the user needs to notice? add Note
}
\note{
For the function to work we need the output of the function get_lake_intersection
that are the lake_reference.txt and intersection table (i.e. coord_lake.txt)
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

data <- fread(paste0(my_directory,
                       "/hydrography90m_test_data",
                       "/coord_lake_1.txt"),
                      header = TRUE)

direction <- (paste0(my_directory,
                    "/hydrography90m_test_data",
                  "/direction_1264942.tif"))

catch <- (paste0(my_directory,
                   "/hydrography90m_test_data/"))

get_lake_catchment(data, direction = direction, catch = catch)
}
\references{
add reference manual html here
https://grass.osgeo.org/grass82/manuals/
add manual of the MSPA analysis tool here
}
\author{
Jaime Garcia Marquez, Thomas Tomiczek
}
