% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Identify the operating system.
The function was written by Will Lowe and was copied from here:
https://conjugateprior.org/2015/06/identifying-the-os-from-r/}
\usage{
get_os()
}
\description{
Identify the operating system.
The function was written by Will Lowe and was copied from here:
https://conjugateprior.org/2015/06/identifying-the-os-from-r/
}
\keyword{internal}
