% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_location_info.R
\name{get_subcID}
\alias{get_subcID}
\title{Join sub-catchment IDs with occurrence points}
\usage{
get_subcID(subc_path, dataset_path)
}
\arguments{
\item{subc_path}{Path to the sub-catchment ID .tif layer}

\item{dataset_path}{Path to data set text file with the lat long columns}
}
\description{
Add the column 'subcID' to the dataset in 'dataset_path' and export the
data frame in a csv.
In the current version the long-lat values must be in the 2nd and 3rd
column of the data set text file.
}
