% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_id.R
\name{get_tile_id}
\alias{get_tile_id}
<<<<<<< HEAD
\title{Identifies the ids of the tiles in which the given points are located. Input is a point data frame.}
=======
\title{Get tile id when the input is a point dataframe}
>>>>>>> dev_get_tile_id
\usage{
get_tile_id(data, lon, lat)
}
\arguments{
\item{data}{data.frame with the lat lon columns}

\item{lon}{Name of the longitude column}

\item{lat}{Name of the latitude column}
}
\description{
<<<<<<< HEAD
Identifies the ids of the tiles in which the given points are located. Input is a point data frame.
=======
Get tile id when the input is a point dataframe
}
\author{
Afroditi Grigoropoulou
>>>>>>> dev_get_tile_id
}
