% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrographr-package.R
\docType{package}
\name{hydrographr-package}
\alias{hydrographr}
\alias{hydrographr-package}
\title{hydrographr: Working with Hydrography90m Data in R}
\description{
A collection of R function wrappers for GDAL and GRASS-GIS functions to efficiently work with Hydrography90m spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glowabio.github.io/hydrographr/}
  \item Report bugs at \url{https://github.com/glowabio/hydrographr/issues}
}

}
\author{
\strong{Maintainer}: Maria Üblacker \email{maria.ueblacker@igb-berlin.de}

Authors:
\itemize{
  \item Afroditi Grigoropoulou \email{afroditi.grigoropoulou@igb-berlin.de}
  \item Jaime Garcia Marquez \email{jaime.marquez@igb-berlin.de}
  \item Yusdiel Torres Cambas \email{yusdiel.torres-cambas@igb-berlin.de}
  \item Christoph Schürz \email{christoph.schuerz@ufz.de}
  \item Mathieu Floury \email{mathieu.floury@univ-lyon1.fr}
  \item Thomas Tomiczek \email{thomas.tomiczek@igb-berlin.de}
  \item Vanessa Bremerich \email{bremerich@igb-berlin.de}
  \item Sami Domisch \email{sami.domisch@igb-berlin.de}
}

}
\keyword{internal}
