% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/utils_download.R
\name{make_sh_exec}
\alias{make_sh_exec}
\title{Make bash scripts executable}
\usage{
make_sh_exec()

make_sh_exec()
}
\description{
Make bash scripts executable

Make bash scripts executable
}
\keyword{internal}
