% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_snapped_points_strahler_async.R
\name{print.ocg_async_job}
\alias{print.ocg_async_job}
\title{Print method for async job objects}
\usage{
\method{print}{ocg_async_job}(x, ...)
}
\description{
Print method for async job objects
}
