% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geopackage.R
\name{read_geopackage}
\alias{read_geopackage}
\title{Read a GeoPackage file}
\usage{
read_geopackage(gpkg, import_as = "data.table", layer_name = NULL)
}
\arguments{
\item{gpkg}{character. Full path of the GeoPackage file.}

\item{import_as}{character. "data.table", "graph", "sf", or "SpatVect".
"data.table" imports data as a data.table. "graph" imports the layer as a
directed graph (igraph object). This option is only possible for a network
layer (e.g. the stream network). "sf" imports the layer as a  spatial data
frame (sf object). "SpatVect" imports the layer as a SpatVector (terra
object). Default is "data.table".}

\item{layer_name}{character. Name of the specific data layer to import from
the GeoPackage. A specific data layer only needs to be defined if the
GeoPackage contains multiple layers. To see the available layers the function
st_layers() from the R package 'sf' can be used. Optional. Default is NULL.}
}
\description{
Reads a GeoPackage vector file from disk either as a table
(data.table), as a directed graph object (igraph), a spatial dataframe (sf)
or a SpatVect object (terra).
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)


# Read the stream network as a graph
my_graph <- read_geopackage(gpkg = paste0(my_directory,
                                          "/hydrography90m_test_data",
                                          "/order_vect_59.gpkg"),
                            import_as = "graph")

# Read the stream network as a data.table
my_dt <- read_geopackage(gpkg = paste0(my_directory,
                                       "/hydrography90m_test_data",
                                       "/order_vect_59.gpkg"))


# Read the sub-catchments as a SF-object
my_sf <- read_geopackage(gpkg = paste0(my_directory,
                                       "/hydrography90m_test_data",
                                       "/sub_catchment_59.gpkg"),
                         import_as = "sf",
                         layer_name = "sub_catchment")

# Read the basin as SpatVect object
my_sv <- read_geopackage(gpkg = paste0(my_directory,
                                       "/hydrography90m_test_data",
                                       "/basin_59.gpkg"),
                         import_as = "SpatVect")

}
\author{
Sami Domisch, Maria M.Üblacker
}
