% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_no_data.R
\name{report_no_data}
\alias{report_no_data}
\title{Report no data value of raster layers.}
\usage{
report_no_data(data_dir, variables, n_cores = NULL)
}
\arguments{
\item{data_dir}{Character. Path to the directory containing all input data.}

\item{variables}{Character vector. The names of the raster files.
e.g. slope_grad_dw_cel_h00v00.tif.
#' @param n_cores Numeric. Number of cores used for parallelization}
}
\description{
Report no data value of raster layers.
}
\author{
Afroditi Grigoropoulou
}
