% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph_utils.R
\name{select_hydro_parameter}
\alias{select_hydro_parameter}
\title{Filter data for hydrometric parameter to determine Y-axis Label}
\usage{
select_hydro_parameter(
  all_hydro_sites_historical,
  all_hydro_sites_1yr,
  parameter
)
}
\arguments{
\item{all_hydro_sites_historical}{A data frame containing historical hydrometric data.}

\item{all_hydro_sites_1yr}{A data frame containing hydrometric data for the current year.}

\item{parameter}{A character string specifying the parameter to filter by. Options are \code{"Flow"} or \code{"Level"}.}
}
\value{
A list with the following components:
\itemize{
\item \code{historical_data}: Filtered historical hydrometric data.
\item \code{current_data}: Filtered current year hydrometric data.
\item \code{y_axis_label}: A character string for the y-axis label, either \code{"Flow (m³/s)"} or \code{"Water Level (m)"}.
}
}
\description{
Filters hydrometric data based on the specified parameter (\code{"Flow"} or \code{"Level"})
and determines the corresponding y-axis label for hydrographs.
}
