% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph_utils.R
\name{set_custom_limit}
\alias{set_custom_limit}
\title{Set Custom Limit for Plotting Axis}
\usage{
set_custom_limit(custom_input, vec1, vec2, multiplier = 1, direction = "max")
}
\arguments{
\item{custom_input}{Numeric. A custom limit input value. If \code{NA}, the function will calculate the limit.}

\item{vec1}{Numeric vector. The first vector to be considered in the limit calculation.}

\item{vec2}{Numeric vector. The second vector to be considered in the limit calculation.}

\item{multiplier}{Numeric. A scaling factor applied to the calculated limit (default is 1).}

\item{direction}{Character. Either "max" to calculate the maximum limit or "min" to calculate the minimum limit. Default is "max".}
}
\value{
Numeric. The calculated axis limit based on the input vectors and scaling factor, or the provided custom input.
If both vectors contain only \code{NA} values, the function returns \code{NA}.
}
\description{
This function calculates a custom limit for the y-axis (or any other axis) by taking
the maximum or minimum of two vectors, with a scaling multiplier. If both vectors
contain only \code{NA} values, the function returns \code{NA}. If a valid custom input is provided,
that value is returned instead of performing the calculation.
}
