% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_no_data.R
\name{set_no_data}
\alias{set_no_data}
\title{Change or set the NoData value for a raster layer.
The change happens in-place, meaning that the original file is overwritten
on disk.}
\usage{
set_no_data(data_dir, variable, no_data)
}
\arguments{
\item{data_dir}{Character. Path to the directory containing all input data.}

\item{variable}{Character vector. Variable file name,
e.g. slope_grad_dw_cel_h00v00.tif. The original file will be overwritten.}

\item{no_data}{The desired NoData value}
}
\description{
Change or set the NoData value for a raster layer.
The change happens in-place, meaning that the original file is overwritten
on disk.
}
\examples{
# Download test data into temporary R folder
my_directory <- tempdir()
download_test_data(my_directory)

# Define no data value
set_no_data(data_dir = paste0(my_directory, "/hydrography90m_test_data"),
            variable = "cti_1264942.tif",
            no_data = -9999)

}
\references{
\url{https://gdal.org/programs/gdal_edit.html}
}
\author{
Afroditi Grigoropoulou, Maria M. Üblacker
}
