% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_to_subc_segment.R
\name{snap_to_subc_segment}
\alias{snap_to_subc_segment}
\title{Snaps data points to the stream segment within the sub-catchment}
\usage{
snap_to_subc_segment(
  data,
  lon,
  lat,
  site_id,
  basin_id = NULL,
  subc_id = NULL,
  basin_path,
  subc_path,
  stream_path,
  cores = 1,
  quiet = TRUE
)
}
\arguments{
\item{data}{Data.frame with lat/lon columns in WGS84.}

\item{lon}{Column name of longitude coordinates as character string.}

\item{lat}{Column name of latitude coordinates as character string.}

\item{site_id}{Column name of unique site IDs as character string. Each row
of the column needs to have an unique ID.}

\item{basin_id}{Column name of basin IDs as character string. If NULL
basin ID will be extracted automatically beforehand.}

\item{subc_id}{Column name of sub-catchment IDs as character string. If
NULL sub-catchment ID will be extracted automatically beforehand.}

\item{basin_path}{Full path of the basin .tif layer.}

\item{subc_path}{Full path of the sub-catchment .tif layer.}

\item{stream_path}{Full path of stream network .gpkg file.}

\item{cores}{Number of cores used for parallelization. By default the process
does not run in parallel.}

\item{quiet}{Whether the standard output will be printed or not.}
}
\description{
Snaps data points to the stream segment of the
sub-catchment the data point is located.
}
\details{
The function uses the network preparation and maintenance module of
GRASS GIS (v.net), to connect a vector lines map (stream network) with a
points map (occurrence/sampling points). After masking the stream segment and
the sub-catchment where the target point is located, the connect operation
snaps the point to the stream segment using a distance threshold. This
threshold is automatically calculated as the longest distance between two
points within the sub-catchment. In this way the snapping will always take
place.This operation creates a new node on the vector line (i.e. stream
segment) from which the new snapped coordinates can be extracted.
}
\examples{
# Download test data into temporary R folder
my_directory <- tempdir()
download_test_data(my_directory)

# Load occurrence data
species_occurence <- read.table(paste0(my_directory,
                            "/hydrography90m_test_data/spdata_1264942.txt"),
                              header = TRUE)
basin_rast <- paste0(my_directory,
                     "/hydrography90m_test_data/basin_1264942.tif")
subc_rast <- paste0(my_directory,
                    "/hydrography90m_test_data/subcatchment_1264942.tif")

# Define full path to the vector file of the stream network
stream_vect <- paste0(my_directory,
                      "/hydrography90m_test_data/order_vect_59.gpkg")

hydrography90m_ids <- extract_ids(data = species_occurence,
                                  lon = "longitude",
                                  lat = "latitude",
                                  site_id = "occurrence_id",
                                  subc_path = subc_rast,
                                  basin_path = basin_rast)

# Snap data points to the stream segment of the provided sub-catchment ID
snapped_coordinates <- snap_to_subc_segment(data = hydrography90m_ids,
                                            lon = "longitude",
                                            lat = "latitude",
                                            site_id = "occurrence_id",
                                            basin_id = "basin_id",
                                            subc_id = "subcatchment_id",
                                            basin_path = basin_rast,
                                            subc_path = subc_rast,
                                            stream_path = stream_vect,
                                            cores = 2)
# Show head of output table
head(snapped_coordinates)

# OR
# Automatically extract the basin and sub-catchment IDs and
# snap the data points to the stream segment
snapped_coordinates <- snap_to_subc_segment(data = species_occurence,
                                            lon = "longitude",
                                            lat = "latitude",
                                            site_id = "occurrence_id",
                                            basin_path = basin_rast,
                                            subc_path = subc_rast,
                                            stream_path = stream_vect,
                                            cores = 2)
# Show head of output table
head(snapped_coordinates)

}
\references{
\url{https://grass.osgeo.org/grass82/manuals/v.net.html}
}
\seealso{
\code{\link{snap_to_network}} to snap the data points to the next stream
segment within a given radius and/or a given flow accumulation threshold
value.
\code{\link{extract_ids}} to extract basin and sub-catchment IDs.
}
\author{
Jaime Garcia Marquez, Maria M. Üblacker
}
