% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_temp_plotting.R
\name{temp_plotting_historical}
\alias{temp_plotting_historical}
\title{Wrangles temperature data and produces a plot with historical and YOI
temperature data.}
\usage{
temp_plotting_historical(climate_data, minYr = NA, maxYr = NA, YOI = NA)
}
\arguments{
\item{climate_data}{a dataframe with the climate information.}

\item{minYr}{The minimum year included in the historical plot calculations.
Leave as NA if you want the minimum year in historical data.}

\item{maxYr}{The maximum year included in the historical plot calculations
Leave as NA if you want the maximum year present in historical data. Also
consider what your YOI is and if you want that included or not.}

\item{YOI}{A single year of interest that will be compared with a
ribbon of historical data.}
}
\value{
A plot with historical temperature stats and a YOI line if indicated.
}
\description{
\code{temp_plotting_historical} produces a historical temperature plot with a
singular year of interest (YOI) highlighted if desired.
}
\examples{
# First, clean raw climate data using clean_climate()

}
