% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_temperature.R
\name{temperature_stats_daily}
\alias{temperature_stats_daily}
\alias{temp_all_years_monthly}
\alias{temp_monthly_mean_historical}
\alias{temp_count_NAs}
\title{Wrangles temperature data}
\usage{
temperature_stats_daily(climate_data, minYr = NA, maxYr = NA)

temp_all_years_monthly(climate_data)

temp_monthly_mean_historical(climate_data, minYr = NA, maxYr = NA)

temp_count_NAs(climate_data)
}
\arguments{
\item{climate_data}{climate data from climate station}

\item{minYr}{The minimum year included in the historical plot calculations.
Leave as NA if you want the minimum year in historical data.}

\item{maxYr}{The maximum year included in the historical plot calculations
Leave as NA if you want the maximum year present in historical data. Also
consider what your YOI is and if you want that included or not.}
}
\value{
dataframe with statistics of air temperature data
}
\description{
\code{temperature_stats_daily} This function calculates temperature stats for the
input length of time on a daily basis.
\code{temp_all_years_monthly} This function gives average temperature stats on a
monthly basis.
\code{temp_count_NAs} This function counts the amount of NA values in a dataset on
an annual basis
}
\examples{
#' # First, clean raw climate data using clean_climate()
#climate_clean <- clean_climate(climate_raw)

# Calculate temperature stats
#temperature_stats_daily(climate_clean, minYr = NA, maxYr = 2023)
#temp_all_years_monthly(climate_clean)
#temp_monthly_mean_historical(climate_clean, minYr = NA, maxYr = 2023)
#temp_count_NAs(climate_clean)

}
