% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_filter_by_attribute.R
\name{api_filter_by_attribute}
\alias{api_filter_by_attribute}
\title{Filter remote CSV using pygeoapi filter-by-attribute process}
\usage{
api_filter_by_attribute(csv_url, keep, comment = NULL)
}
\arguments{
\item{csv_url}{Character. URL of the CSV file to filter.}

\item{keep}{Named list. Each element name is a column in the CSV and each
element value is a vector of values to retain. Example:
`list(site_id = c("FP1", "FP10"))`.}

\item{comment}{Optional comment that will be stored in the API response.}
}
\value{
A character string with the `href` pointing to the filtered CSV.
}
\description{
Sends a POST request to the pygeoapi `filter-by-attribute` process,
providing a remote CSV URL and a set of attribute filters, and
returns the URL (`href`) to the filtered output CSV generated by the API.
}
\examples{
\dontrun{
api_filter_by_attribute(
  csv_url = "https://aqua.igb-berlin.de/referencedata/aqua90m/spdata_species.csv",
  keep = list(site_id = c("FP1", "FP2"))
)
}
}
\seealso{
Other ocgapi: 
\code{\link{api_get_basin_streamsegments}()},
\code{\link{api_get_env90m_data}()},
\code{\link{api_get_local_ids}()},
\code{\link{api_get_snapped_points}()},
\code{\link{api_get_snapped_points_cascade_plural}()},
\code{\link{api_get_snapped_points_df}()},
\code{\link{api_get_snapped_points_strahler}()},
\code{\link{api_get_snapped_points_strahler_plural}()},
\code{\link{api_get_upstream_subcids}()}
}
\concept{ocgapi}
