% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_basin_stream_segments.R
\name{api_get_basin_streamsegments}
\alias{api_get_basin_streamsegments}
\title{Get Stream Segments for a Basin}
\usage{
api_get_basin_streamsegments(
  basin_id = NULL,
  subc_id = NULL,
  geometry_only = FALSE,
  comment = NULL,
  strahler_min = NULL
)
}
\arguments{
\item{basin_id}{Integer. The ID of the basin to retrieve stream segments for.}

\item{geometry_only}{Logical. If `TRUE`, returns only the geometry of the stream
segments without associated attributes. Defaults to `FALSE`.}

\item{strahler_min}{Integer (optional). The minimum Strahler stream order to include.
Streams below this order will be excluded. Defaults to `NULL` (no filtering).}
}
\value{
An `sf` object representing the stream segments of the given basin.
}
\description{
This function returns all stream segments within a specified basin as spatial
features, optionally filtering by a minimum Strahler order.
}
\details{
Sends a request to the GeoFRESH API to retrieve
stream segments belonging to a given basin and returns them as an `sf` object.
}
\examples{
\dontrun{
segm_sf <- api_get_basin_streamsegments(
  basin_id = 1288419,
  geometry_only = FALSE,
  strahler_min = 3
)

library(leaflet)
leaflet(segm_sf) |>
  addProviderTiles("CartoDB.Positron") |>
  addPolylines(color = "blue")
}

}
\seealso{
Other ocgapi: 
\code{\link{api_filter_by_attribute}()},
\code{\link{api_get_env90m_data}()},
\code{\link{api_get_local_ids}()},
\code{\link{api_get_snapped_points}()},
\code{\link{api_get_snapped_points_cascade_plural}()},
\code{\link{api_get_snapped_points_df}()},
\code{\link{api_get_snapped_points_strahler_plural}()},
\code{\link{api_get_upstream_subcids}()}
}
\concept{ocgapi}
