% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_job_results.R
\name{api_get_job_results}
\alias{api_get_job_results}
\title{Download and return results from a completed job}
\usage{
api_get_job_results(
  jobID,
  base_url = "https://aqua.igb-berlin.de/pygeoapi-dev",
  as_dataframe = TRUE
)
}
\arguments{
\item{jobID}{Character. Job ID or the result object from api_poll_job().}

\item{base_url}{Character. Base URL for the pygeoapi instance.}

\item{as_dataframe}{Logical. If TRUE (default), return dataframe. If FALSE, return URL only.}
}
\value{
Data frame with the results, or character URL if as_dataframe = FALSE
}
\description{
Download and return results from a completed job
}
