% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_local_ids.R
\name{api_get_local_ids}
\alias{api_get_local_ids}
\title{Retrieve Local IDs for Sampling Sites}
\usage{
api_get_local_ids(
  csv_url,
  colname_lat = "latitude",
  colname_lon = "longitude",
  colname_site_id = "site_id",
  colname_subc_id = NULL,
  which_ids = "reg_id",
  comment = NULL
)
}
\arguments{
\item{csv_url}{URL to a CSV file hosted online.}

\item{colname_lat}{Name of the latitude column in CSV file.}

\item{colname_lon}{Name of the longitude column in CSV file.}

\item{colname_site_id}{Name of the site ID column in CSV file.}

\item{colname_subc_id}{Name of the subcatchment ID column in CSV file (optional).}

\item{which_ids}{Which ID type(s) to retrieve, e.g. `"reg_id"`, `"basin_id"`,
`"subc_id"`.}

\item{comment}{Optional comment string.}
}
\value{
A list with two elements:
\describe{
  \item{data}{A `data.frame` with the retrieved IDs.}
  \item{href}{A `character` string with the URL of the output CSV.}
}
}
\description{
Queries the GeoFRESH API to obtain regional unit, basin,
and subcatchment IDs for a set of sampling locations.
Handles both single-point and multi-point inputs.
}
\seealso{
Other ocgapi: 
\code{\link{api_filter_by_attribute}()},
\code{\link{api_get_basin_streamsegments}()},
\code{\link{api_get_env90m_data}()},
\code{\link{api_get_snapped_points}()},
\code{\link{api_get_snapped_points_cascade_plural}()},
\code{\link{api_get_snapped_points_df}()},
\code{\link{api_get_snapped_points_strahler}()},
\code{\link{api_get_snapped_points_strahler_plural}()},
\code{\link{api_get_upstream_subcids}()}
}
\author{
Afroditi Grigoropoulou, Merret Buurman
}
\concept{ocgapi}
