% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_snapped_points_strahler.R
\name{api_get_snapped_points_strahler}
\alias{api_get_snapped_points_strahler}
\title{Snap Points to River Network Using Strahler Threshold}
\usage{
api_get_snapped_points_strahler(
  df,
  colname_lat = "latitude",
  colname_lon = "longitude",
  colname_site_id = "site_id",
  strahler = NULL,
  geometry_only = FALSE,
  comment = NULL
)
}
\arguments{
\item{df}{A data.frame containing point data.}

\item{colname_lat}{Name of the latitude column. Default: "latitude".}

\item{colname_lon}{Name of the longitude column. Default: "longitude".}

\item{colname_site_id}{Name of the site ID column. Default: "site_id".}

\item{strahler}{Minimum Strahler order to snap to (e.g., 3).}

\item{geometry_only}{Logical. If TRUE, return only geometry.}

\item{comment}{Optional comment string passed to the API.}

\item{process_url}{URL of the pygeoapi process (e.g., "http://localhost:5000/processes/get-snapped-points-strahler/execution").}
}
\value{
A data.frame with the snapped point data.
}
\description{
Sends individual coordinate points to the GeoFRESH API to snap them
to the nearest stream segment above a given Strahler order.
}
\examples{
\dontrun{
# Example usage
url <- "https://example.com/sites.csv"
df = data.frame(site_id = c(1,2),
          longitude = c(21.02478, 20.89024),
          latitude =  c(40.10572, 40.14585))
 snapped_df <- api_get_snapped_points_strahler(df = df,
  colname_lat = "latitude",
  colname_lon = "longitude",
  colname_site_id = "site_id",
  strahler = 4)

#' head(result$data)
cat("Download link:", result$href)

}


}
\seealso{
Other ocgapi: 
\code{\link{api_filter_by_attribute}()},
\code{\link{api_get_basin_streamsegments}()},
\code{\link{api_get_env90m_data}()},
\code{\link{api_get_local_ids}()},
\code{\link{api_get_snapped_points}()},
\code{\link{api_get_snapped_points_cascade_plural}()},
\code{\link{api_get_snapped_points_df}()},
\code{\link{api_get_snapped_points_strahler_plural}()},
\code{\link{api_get_upstream_subcids}()}
}
\author{
Afroditi Grigoropoulou, Merret Buurman
}
\concept{ocgapi}
