% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tiles_filesize.R
\name{check_tiles_filesize}
\alias{check_tiles_filesize}
\title{Check the size of single files before downloading.
It is called and inherits arguments by the function 'download_tiles()'}
\usage{
check_tiles_filesize(
  variable,
  file_format = "tif",
  tile_id = NULL,
  reg_unit_id = NULL,
  global = FALSE,
  h90m_varnames,
  h90m_tile_id,
  h90m_file_formats,
  file_size_table_sep
)
}
\arguments{
\item{variable}{character vector of variable names.}

\item{file_format}{character. Format of the requested file ("tif" or "gpkg").}

\item{tile_id}{character. The ID of the requested tile or regional unit.}

\item{global}{logical. If TRUE, the global extent file is downloaded.
Default is FALSE.}

\item{h90m_varnames}{character vector. the valid names of the hydrography90m
files available for download, (inherited by 'download_tiles()').}

\item{h90m_tile_id}{character vector. The valid IDs of the hydrography90m.
regular tiles available for download (inherited by 'download_tiles()').}

\item{h90m_file_formats}{character vector. The valid file types of the files
available for download (inherited by 'download_tiles()').}

\item{file_size_table_sep}{data.frame. Lookup table including file names
and sizes (inherited by 'download_tiles()').}
}
\description{
Check the size of single files before downloading.
It is called and inherits arguments by the function 'download_tiles()'
}
\keyword{internal}
