% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_produce_hydrographs.R
\name{create_hydrograph_separate}
\alias{create_hydrograph_separate}
\alias{create_hydrograph_faceted}
\title{Create Hydrograph Plots}
\usage{
create_hydrograph_separate(
  all_hydro_sites_historical,
  all_hydro_sites_1yr,
  parameter,
  output_type = "print",
  WY = FALSE,
  custom_ymax_input = NA,
  custom_ymin_input = NA
)

create_hydrograph_faceted(
  all_hydro_sites_historical,
  all_hydro_sites_1yr,
  parameter = "flow",
  WY = FALSE,
  output_type = "print",
  fixed_y_scales = "fixed",
  custom_ymax_input = NA,
  custom_ymin_input = NA,
  jpeg_width = 6,
  jpeg_height = 8
)
}
\arguments{
\item{all_hydro_sites_historical}{A data frame containing historical hydrology data.}

\item{all_hydro_sites_1yr}{A data frame containing one-year hydrology data.}

\item{parameter}{A character string specifying the parameter to plot. Options are \code{"flow"} or \code{"level"}.}

\item{output_type}{A character string indicating the output format. Options are \code{"JPEG"} for JPEG figure output
or \code{"print"} for HTML/Markdown output.}

\item{WY}{Logical value indicating whether to present hydrograph by water year (\code{TRUE}) or calendar year (\code{FALSE}).}

\item{custom_ymax_input}{A numeric value for a custom maximum y-axis value.}

\item{custom_ymin_input}{A numeric value for a custom minimum y-axis value.}

\item{fixed_y_scales}{A character string specifying whether the y-axis scale is \code{"fixed"} or \code{"free"}
across facets. Defaults to \code{"fixed"}.}

\item{jpeg_width}{A numeric value specifying the width of the figure (in pixels) for JPEG output.}

\item{jpeg_height}{A numeric value specifying the height of the figure (in pixels) for JPEG output.}
}
\value{
Hydrograph plots saved to the \code{"figures/"} folder.
}
\description{
\itemize{
\item \strong{\code{create_hydrograph_separate}}: Generates separate hydrograph plots for specified locations.
\item \strong{\code{create_hydrograph_faceted}}: Creates a joint hydrograph plot for specified locations using \code{facet_grid}.
}
}
\examples{
# Download hydrology data for multiple stations
hydro_data <- dl_hydro(station_number = c("08LD001", "08LC002"))

# Generate historical and single-year hydrology statistics
all_hydro_sites_hist <- create_hydro_stats_historical(hydro_data)
all_hydro_sites_1yr <- create_hydro_stats_singleYr(hydro_data, YOI = 2021, WY = TRUE)

# Create separate hydrograph plots
create_hydrograph_separate(all_hydro_sites_hist, all_hydro_sites_1yr, 
                           parameter = "flow", WY = TRUE)

# Create faceted hydrograph plot
create_hydrograph_faceted(all_hydro_sites_hist, all_hydro_sites_1yr)

}
