% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_to_extent.R
\name{crop_to_extent}
\alias{crop_to_extent}
\title{crop_to_extent}
\usage{
crop_to_extent(
  raster_path,
  vector_path = NULL,
  bound_box = NULL,
  output_path,
  rcrop_read = TRUE
)
}
\arguments{
\item{raster_path}{Path to the raster .tif layer}

\item{vector_path}{Path to a vector layer that is used as a cutline data
source}

\item{bound_box}{The coordinates of the corner of a bounding box (ULX: upper left x value
bounding box, ULY: upper left y value bounding box, LRX: lower right x value
bounding box, LRY: lower right y value bounding box).}

\item{output_path}{Path for the raster .tif output}

\item{rcrop_read}{If TRUE the cropped raster .tif layer gets read into R.
TRUE is set by default.}
}
\value{
A .tif raster file and a SpatRaster object, or a .tif raster file if
rcrop_read = TRUE
}
\description{
Crop a raster .tif to a polygon border line if a vector layer
is provided, otherwise if the coordinates of a bounding box are provided,
the raster is cropped to the extent of the bounding box. At least a cutline
source (vector layer) or a bounding box coordinates must be provided.
}
