% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_to_extent.R
\name{crop_to_extent}
\alias{crop_to_extent}
\title{crop_to_extent}
\usage{
crop_to_extent(
  raster_path,
  vector_path = NULL,
  bound_box = NULL,
  output_path,
  rcrop_read = TRUE
)
}
\arguments{
\item{raster_path}{Path to the raster .tif layer}

\item{vector_path}{Path to a vector layer that is used as a cutline data
source}

\item{bound_box}{The coordinates of the corner of a bounding box (xmin, ymin,
xmax, ymax) or a SpatRaster, SpatVector, or other spatial object.}

\item{output_path}{Path for the raster .tif output}

\item{rcrop_read}{If TRUE the cropped raster .tif layer gets read into R.
TRUE is set by default.}
}
\value{
A .tif raster file and a SpatRaster object, or a .tif raster file if
rcrop_read = FALSE
}
\description{
Crop a raster .tif to a polygon border line if a vector layer
(cutline source) is provided, otherwise if a bounding box is provided
(xmin, ymin, xmax, ymax coordinates or a spatial object from which to extract
 a bounding box), the raster is cropped to the extent of the bounding box. At
 least a cutline source (vector_path) or a bounding box (bound_box)
 must be provided. If rcrop_read = TRUE (default), the outputs are a .tif
 (saved under output_path) and a SpatRaster (terra package) object,
 otherwise if rcrop_read = FALSE, the .tif file is the only output.
}
