% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_to_extent.R
\name{crop_to_extent}
\alias{crop_to_extent}
\title{Crops a raster layer to a specific extent on disk}
\usage{
crop_to_extent(
  raster_layer,
  vector_layer = NULL,
  bounding_box = NULL,
  out_dir,
  file_name,
  read = TRUE
)
}
\arguments{
\item{raster_layer}{character. Full path to the input raster .tif layer.}

\item{vector_layer}{character. Full path to a vector layer that is used as a
cutline data source (similar to a mask operation).}

\item{bounding_box}{numeric vector of the coordinates of the corners of a
bounding box (xmin, ymin, xmax, ymax), SpatRaster, SpatVector,
or other spatial object.}

\item{out_dir}{character. The directory where the output will be stored.}

\item{file_name}{character. Name of the cropped output raster .tif file.}

\item{read}{logical. If TRUE, the cropped raster .tif layer gets read into R.
If FALSE, the layer is only stored on disk. Default is TRUE.}
}
\value{
The function returns always a .tif raster file written to disk.
Optionally, a SpatRaster (terra object) can be loaded into R with
read = TRUE.
}
\description{
This function crops an input raster layer directly on disk,
i.e. the input layer does not need to be loaded into R. The raster .tif is
cropped to a polygon border line if a vector layer (cutline source) is
provided, otherwise if a bounding box is provided
(xmin, ymin, xmax, ymax coordinates or a spatial object from which to extract
a bounding box), the raster is cropped to the extent of the bounding box. At
least a cutline source (vector_layer) or a bounding box (bounding_box)
should be provided. The output is always written to disk, and can be
optionally loaded into R as a SpatRaster (terra package) object
(using read = TRUE).
}
\examples{
library(hydrographr)
library(data.table)

# Specify the working directory of the test data
DATADIR <- "path/to/test_data"

# Download the test data
download_test_data(DATADIR)

# Crop the Stream Power Index to the basin
spi_basin <- crop_to_extent(raster_layer = paste0(DATADIR,
"/spi_1264942.tif"),
vector_layer = paste0(DATADIR, "/basin_59.gpkg"),
output_path = "/my/output/path/spi_basin_cropped.tif"),
read = TRUE)
}
\author{
Yusdiel Torres-Cambas
}
