% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_data.R
\name{download_test_data}
\alias{download_test_data}
\title{Download test data}
\usage{
download_test_data(download_dir = ".")
}
\arguments{
\item{download_dir}{character. The directory where the files will be
downloaded. Default location is the working directory.}
}
\description{
Download the test data of the package, which includes all
Hydrography90m and species point observation data
for a small geographic extent, to test the functions.

The test data is available at
https://drive.google.com/file/d/1kYNWXmtVm6X7MZLISOePGpvxB1pk1scD/view?usp=share_link
and can be automatically downloaded and unzipped
with this function to a desired path.
}
\details{
Downloads the test data for the Hydrography90m dataset
}
\examples{
# Download the test data to the current working directory
download_test_data()

# Download the data to a specific (existing) directory
download_test_data("path/to/your/directory")

}
\references{
Amatulli, G., Garcia Marquez, J., Sethi, T., Kiesel, J., Grigoropoulou, A.,
Üblacker, M. M., Shen, L. Q., and Domisch, S.: Hydrography90m: a new
high-resolution global hydrographic dataset, Earth Syst. Sci. Data, 14,
4525–4550, https://doi.org/10.5194/essd-14-4525-2022, 2022.")

Amatulli G., Garcia Marquez J., Sethi T., Kiesel J., Grigoropoulou A.,
Üblacker M., Shen L. & Domisch S. (2022-08-09 ). Hydrography90m: A new
high-resolution global hydrographic dataset. IGB Leibniz-Institute of
Freshwater Ecology and Inland Fisheries. dataset.
https://doi.org/10.18728/igb-fred-762.1
}
\author{
Afroditi Grigoropoulou
}
