% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_data.R
\name{download_test_data}
\alias{download_test_data}
\title{Downloads all the test data from https://drive.google.com/drive/u/1/folders/18RpqmywL7yThitsOklWXAuYIfMvfyPfg.}
\usage{
download_test_data(download_path = "./hydrography90m_test_data.zip")
}
\arguments{
\item{download_path}{The path where the files will be downloaded}
}
\description{
Downloads all the test data from https://drive.google.com/drive/u/1/folders/18RpqmywL7yThitsOklWXAuYIfMvfyPfg.
}
\author{
Afroditi Grigoropoulou
}
