% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_data.R
\name{download_test_data}
\alias{download_test_data}
\title{Downloads the test data for the Hydrography90m dataset}
\usage{
<<<<<<< HEAD
download_test_data(download_path = ".")
=======
download_test_data(download_path = "./hydrography90m_test_data")
>>>>>>> dev_download_tiles
}
\arguments{
\item{download_path}{The path where the files will be downloaded}
}
\description{
The test data contains all Hydrography90m and species point observation data for a small geographic extent to test the functions.
}
\details{
The test data is available at https://drive.google.com/file/d/1kYNWXmtVm6X7MZLISOePGpvxB1pk1scD/view?usp=share_link and can be automatically downloaded and unzipped with this function to a desired path.
}
\examples{
# Download the test data to the current working directory
download_test_data()

<<<<<<< HEAD
# Download the data to a specific directory
=======
# Download the data to a specific (existing) directory
>>>>>>> dev_download_tiles
download_test_data("path/to/your/directory")
}
\author{
Afroditi Grigoropoulou
}
