% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tiles.R
\name{download_tiles}
\alias{download_tiles}
\title{Downloads multiple files from Nimbus
by calling the function download_tiles_base in a loop.}
\usage{
download_tiles(
  variable,
  file_format = "tif",
  tile_id = NULL,
  reg_unit_id = NULL,
  global = FALSE,
  download_dir = "."
)
}
\arguments{
\item{variable}{character vector of variable names.}

\item{file_format}{character. Format of the requested file ("tif" or "gpkg").}

\item{tile_id}{character vector. The IDs of the requested tiles.}

\item{reg_unit_id}{character vector. The IDs of the requested regional units.}

\item{global}{logical. If TRUE, the global extent file is downloaded.
Default is FALSE.}

\item{download_dir}{character. The directory where the files will be
downloaded. Default is the working directory.}
}
\description{
Downloads multiple files from Nimbus
by calling the function download_tiles_base in a loop.
}
\examples{
# Download data for two variables in three regular tiles
# to the current working directory
download_tiles(variable = c("sti", "stream_dist_up_farth"),
file_format = "tif",
tile_id = c("h00v02","h16v02", "h16v04"))

# Download the global .tif layer for the variable "direction"
# to a specific (existing) directory
download_tiles(variable = "direction",
file_format = "tif",
global = TRUE,
download_dir = "/path/to/directory/")

}
\references{
Amatulli G., Garcia Marquez J., Sethi T., Kiesel J.,
Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 )
Hydrography90m: A new high-resolution global hydrographic dataset.
IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries.
dataset. https://doi.org/10.18728/igb-fred-762.1
}
