% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tiles.R
\name{download_tiles}
\alias{download_tiles}
\title{Downloads multiple files from Nimbus
by calling the function download_tiles_base in a loop.}
\usage{
download_tiles(
  variable,
  filetype = "tif",
  tile_id = NULL,
  reg_unit_id = NULL,
  global = FALSE,
  download_path = "."
)
}
\arguments{
\item{variable}{vector of variable names (character)}

\item{filetype}{format of the requested file ("tif" or "gpkg")}

\item{tile_id}{id of the requested tile (character)}

\item{reg_unit_id}{id of the requested regional unit (character)}

\item{global}{Should the global file be downloaded or not.
TRUE/FALSE, FALSE by default}

\item{download_path}{The path where the files will be downloaded}
}
\description{
Downloads multiple files from Nimbus
by calling the function download_tiles_base in a loop.
}
\references{
Amatulli G., Garcia Marquez J., Sethi T., Kiesel J.,
Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 )
Hydrography90m: A new high-resolution global hydrographic dataset.
IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries.
dataset. https://doi.org/10.18728/igb-fred-762.1
}
