% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tiles.R
\name{download_tiles}
\alias{download_tiles}
\title{Download files of the Hydrography90m dataset}
\usage{
download_tiles(
  variable,
  file_format = "tif",
  tile_id = NULL,
  reg_unit_id = NULL,
  global = FALSE,
  download_dir = "."
)
}
\arguments{
\item{variable}{character vector of variable names.}

\item{file_format}{character. Format of the requested file ("tif" or "gpkg").}

\item{tile_id}{character vector. The IDs of the requested tiles.}

\item{reg_unit_id}{character vector. The IDs of the requested regional units.}

\item{global}{logical. If TRUE, the global extent file is downloaded.
Default is FALSE.}

\item{download_dir}{character. The directory where the files will be
downloaded. Default is the working directory.}
}
\description{
The function downloads files of the Hydrography90m
dataset, available at https://public.igb-berlin.de/index.php/s/agciopgzXjWswF4?path=%2F .
The files will be stored in the folder architecture of the above domain.
Multiple regular tile or regional unit files can be requested in a single
call of the function. The tile or regional unit IDs can be obtained
using the functions "get_tile_id" and "get_regional_unit_id" respectively.
}
\examples{
# Download data for two variables in three regular tiles
# to the current working directory
download_tiles(variable = c("sti", "stream_dist_up_farth"),
               file_format = "tif",
               tile_id = c("h00v02","h16v02", "h16v04"))

# Download the global .tif layer for the variable "direction"
# to a specific (existing) directory
# Define directory
my_directory <- tempdir()
# Download layer
download_tiles(variable = "direction",
               file_format = "tif",
               global = TRUE,
               download_dir = my_directory)

}
\references{
Amatulli G., Garcia Marquez J., Sethi T., Kiesel J.,
Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 )
Hydrography90m: A new high-resolution global hydrographic dataset.
IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries.
dataset. https://doi.org/10.18728/igb-fred-762.1
}
\author{
Afroditi Grigoropoulou
}
