% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tiles.R
\name{download_tiles}
\alias{download_tiles}
\title{Download files of the Hydrography90m dataset}
\usage{
download_tiles(
  variable,
  file_format = "tif",
  tile_id = NULL,
  reg_unit_id = NULL,
  global = FALSE,
  download_dir = "."
)
}
\arguments{
\item{variable}{character vector of variable names. See Details for all the
variable names.}

\item{file_format}{character. Format of the requested file ("tif" or "gpkg").
See Details.}

\item{tile_id}{character vector. The IDs of the requested tiles.}

\item{reg_unit_id}{character vector. The IDs of the requested regional units.}

\item{global}{logical. If TRUE, the global extent file is downloaded.
Default is FALSE.}

\item{download_dir}{character. The directory where the files will be
downloaded. Default is the working directory.}
}
\description{
The function downloads files of the Hydrography90m
dataset, available at https://public.igb-berlin.de/index.php/s/agciopgzXjWswF4?path=\%2F .
The files will be stored in the folder architecture of the above domain.
Multiple regular tile or regional unit files can be requested in a single
call of the function. The tile or regional unit IDs can be obtained
using the functions "get_tile_id" and "get_regional_unit_id" respectively.
}
\details{
In the following table you can find all the variables included in the
Hydrography90m dataset. The column "Abbreviation" includes the variable names
that should be used as an input in the parameter "variable" of the function.
Likewise, the column "File format" contains the input that should be given to
the "file_format" parameter.
For more details and visualisations of the spatial layers, please refer to
\url{https://hydrography.org/hydrography90m/hydrography90m_layers/}.\tabular{lllll}{
   \strong{Variable type} \tab \strong{Variable} \tab \strong{Abbreviation} \tab \strong{Unit} \tab \strong{File_format} \cr
   Network \tab Drainage basin \tab basin \tab  \tab tif \cr
   Network \tab Drainage basin \tab basin \tab  \tab gpkg \cr
   Network \tab Sub-catchment \tab sub_catchment \tab  \tab tif \cr
   Network \tab Sub-catchment \tab sub_catchment \tab  \tab gpkg \cr
   Network \tab Stream segment \tab segment \tab  \tab tif \cr
   Network \tab Outlet \tab outlet \tab  \tab tif \cr
   Network \tab Outlet \tab outlet \tab  \tab gpkg \cr
   Network \tab Regional unit \tab regional_unit \tab  \tab tif \cr
   Flow \tab Flow accumulation \tab flow \tab km^2 \tab tif \cr
   Stream slope \tab Cell maximum curvature \tab slope_curv_max_dw_cel \tab 1/m \tab tif \cr
   Stream slope \tab Cell minimum curvature \tab slope_curv_min_dw_cel \tab 1/m \tab tif \cr
   Stream slope \tab Cell elevation difference \tab slope_elv_dw_cel \tab m \tab tif \cr
   Stream slope \tab Cell gradient \tab slope_grad_dw_cel \tab  \tab tif \cr
   Stream distance \tab Shortest distance to drainage divide \tab stream_dist_up_near \tab m \tab tif \cr
   Stream distance \tab Longest distance to drainage divide \tab stream_dist_up_farth \tab m \tab tif \cr
   Stream distance \tab Nearest down stream stream grid cell \tab stream_dist_dw_near \tab m \tab tif \cr
   Stream distance \tab Outlet grid cell in the network \tab outlet_dist_dw_basin \tab m \tab tif \cr
   Stream distance \tab Down stream stream node grid cell \tab outlet_dist_dw_scatch \tab m \tab tif \cr
   Stream distance \tab Euclidean distance \tab stream_dist_proximity \tab m \tab tif \cr
   Elevation difference \tab Shortest path \tab stream_diff_up_near \tab m \tab tif \cr
   Elevation difference \tab Longest path \tab stream_diff_up_farth \tab m \tab tif \cr
   Elevation difference \tab Nearest downstream stream pixel \tab stream_diff_dw_near \tab m \tab tif \cr
   Elevation difference \tab Outlet grid cell in the network \tab outlet_diff_dw_basin \tab m \tab tif \cr
   Elevation difference \tab Downstream stream node grid cell \tab outlet_diff_dw_scatch \tab m \tab tif \cr
   Segment properties \tab Segment downstream mean gradient \tab channel_grad_dw_seg \tab  \tab tif \cr
   Segment properties \tab Segment upstream mean gradient \tab channel_grad_up_seg \tab  \tab tif \cr
   Segment properties \tab Cell upstream gradient \tab channel_grad_up_cel \tab  \tab tif \cr
   Segment properties \tab Cell stream course curvature \tab channel curv_cel \tab  \tab tif \cr
   Segment properties \tab Segment downstream elevation difference \tab channel_elv_dw_seg \tab  \tab tif \cr
   Segment properties \tab Segment upstream elevation difference \tab channel_elv_up_seg \tab  \tab tif \cr
   Segment properties \tab Cell upstream elevation difference \tab channel_elv_up_cel \tab  \tab tif \cr
   Segment properties \tab Cell downstream elevation difference \tab channel_elv_dw_cel \tab  \tab tif \cr
   Segment properties \tab Segment downstream distance \tab channel_dist_dw_seg \tab  \tab tif \cr
   Segment properties \tab Segment upstream distance \tab channel_dist_up_seg \tab  \tab tif \cr
   Segment properties \tab Cell upstream distance \tab channel_dist_up_cel \tab  \tab tif \cr
   Stream order \tab Strahler’s stream order \tab order_strahler \tab  \tab tif \cr
   Stream order \tab Shreve’s stream magnitude \tab order_shreve \tab  \tab tif \cr
   Stream order \tab Horton’s stream order \tab order_horton \tab  \tab tif \cr
   Stream order \tab Hack’s stream order \tab order_hack \tab  \tab tif \cr
   Stream order \tab Topological dimension of streams \tab order_topo \tab  \tab tif \cr
   Stream order \tab Strahler’s stream order \tab order_vect_segment \tab  \tab gpkg \cr
   Stream order \tab Shreve’s stream magnitude \tab order_vect_segment \tab  \tab gpkg \cr
   Stream order \tab Horton’s stream order \tab order_vect_segment \tab  \tab gpkg \cr
   Stream order \tab Hack’s stream order \tab order_vect_segment \tab  \tab gpkg \cr
   Stream order \tab Topological dimension of streams \tab order_vect_segment \tab  \tab gpkg \cr
   Stream reach \tab Length of the stream reach \tab order_vect_segment \tab m \tab gpkg \cr
   Stream reach \tab Straight length \tab order_vect_segment \tab m \tab gpkg \cr
   Stream reach \tab Sinusoid of the stream reach \tab order_vect_segment \tab  \tab gpkg \cr
   Stream reach \tab Accumulated length \tab order_vect_segment \tab m \tab gpkg \cr
   Stream reach \tab Flow accumulation \tab order_vect_segment \tab km^2 \tab gpkg \cr
   Stream reach \tab Distance to outlet \tab order_vect_segment \tab m \tab gpkg \cr
   Stream reach \tab Source elevation \tab order_vect_segment \tab m \tab gpkg \cr
   Stream reach \tab Outlet elevation \tab order_vect_segment \tab m \tab gpkg \cr
   Stream reach \tab Elevation drop \tab order_vect_segment \tab  \tab gpkg \cr
   Stream reach \tab Outlet drop \tab order_vect_segment \tab  \tab gpkg \cr
   Stream reach \tab Gradient \tab order_vect_segment \tab  \tab gpkg \cr
   Flow index \tab Stream power index \tab spi \tab  \tab tif \cr
   Flow index \tab Sediment transportation index \tab sti \tab  \tab tif \cr
   Flow index \tab Compound topographic index \tab cti \tab  \tab tif \cr
}
}
\examples{
# Download data for two variables in three regular tiles
# to the current working directory
download_tiles(variable = c("sti", "stream_dist_up_farth"),
               file_format = "tif",
               tile_id = c("h00v02","h16v02", "h16v04"))

# Download the global .tif layer for the variable "direction"
# into the temporary R folder or define a different directory
# Define directory
my_directory <- tempdir()
# Download layer
download_tiles(variable = "direction",
               file_format = "tif",
               global = TRUE,
               download_dir = my_directory)

}
\references{
Amatulli G., Garcia Marquez J., Sethi T., Kiesel J.,
Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 )
Hydrography90m: A new high-resolution global hydrographic dataset.
IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries.
dataset. https://doi.org/10.18728/igb-fred-762.1
}
\author{
Afroditi Grigoropoulou
}
