% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_gpkg.R
\name{extract_from_gpkg}
\alias{extract_from_gpkg}
\title{Extract values from the stream order .gpkg files.}
\usage{
extract_from_gpkg(
  data_dir,
  out_path = NULL,
  subc_ids,
  subc_layer,
  variables,
  n_cores = NULL
)
}
\arguments{
\item{data_dir}{Character. Path to the directory containing all input data}

\item{out_path}{Character. Full path of the output file.
If not NULL, the output data.frame is exported as a csv in the given path}

\item{subc_ids}{Vector of sub-catchment ids or "all".
If "all", the gpkg values are extracted for all the segments
of the given gpkg layer.
The stream segment ids are the same as the sub-catchment ids.
A vector of the sub-catchment ids can be acquired
from the extract_ids() function,
by sub setting the resulting data.frame}

\item{subc_layer}{Character. Full path to the sub-catchment ID .tif layer}

\item{variables}{Character vector. Variable file names,
e.g. order_vect_point_h18v04.gpkg.
Variable names should remain intact in file names,
even after prior file processing,
i.e., order_vect_point should appear in the file name.
The files should be cropped to the extent of the sub-catchment layer}

\item{n_cores}{Numeric. Number of cores for parallelization}
}
\description{
Extract values from the stream order .gpkg files.
}
\author{
Afroditi Grigoropoulou Jaime Garcia Marquez
}
