% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ids.R
\name{extract_ids}
\alias{extract_ids}
\title{Extract sub-catchment and/or basin IDs of a point location}
\usage{
extract_ids(
  data,
  lon,
  lat,
  site_id = NULL,
  basin_path = NULL,
  subc_path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{data}{Data.frame with lat/lon coordinates in WGS84.}

\item{lon}{Column name of longitude coordinates as character vector.}

\item{lat}{Column name of latitude coordinates as character vector.}

\item{site_id}{Column name of a unique ID as character vector; Optional
an ID for the point location can be defined.}

\item{basin_path}{Full path to the basin ID .tif layer.}

\item{subc_path}{Full path to the sub-catchment ID .tif layer.}

\item{quiet}{Whether the standard output will be printed or not.}
}
\description{
Extracts the ID value of the basin and/or sub-catchment raster layer at a
given point location.
}
\details{
For the extraction of a value at a given point location from the basin
and/or sub-catchment raster layer of the Hydrography90m the GDAL function
'gdallocationinfo' is used. The point locations have to be defined by
coordinates of the WGS84 reference system, it can be used to extract any
value from a given raster layer with a WGS84 projection.
}
\note{
Duplicated rows will be removed.
}
\examples{
# Download test data into temporary R folder
download_test_data(tempdir())

# Load occurrence data
species_occurrence <- read.table(
paste0(tempdir(), "/hydrography90m_test_data/spdata_1264942.txt"),
header = TRUE)

# Define full path to the basin and sub-catchments raster layer
basin_rast <- paste0(tempdir(),
                     "/hydrography90m_test_data/basin_1264942.tif")
subc_rast <- paste0(tempdir(),
                    "/hydrography90m_test_data/basin_1264942.tif")

# Extract basin and sub-catchment IDs from the Hydrography90m layers
hydrography90m_ids <- extract_ids(data = species_occurrence,
                                  lon = "longitude",
                                  lat = "latitude",
                                  site_id = "occurrence_id",
                                  subc_path = subc_rast,
                                  basin_path = basin_rast)

# Show head of the output table
head(hydrography90m_ids)
}
\references{
\url{https://gdal.org/programs/gdallocationinfo.html}
}
\author{
Afroditi Grigoropoulou, Maria Üblacker
}
