% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ids.R
\name{extract_ids}
\alias{extract_ids}
\title{Join sub-catchment IDs with occurrence points}
\usage{
extract_ids(dataset_path, lon, lat, subc_path, basin_path)
}
\arguments{
\item{dataset_path}{Path to data set text file with the lat long columns}

\item{lon}{Name of the longitude column}

\item{lat}{Name of the latitude column}

\item{subc_path}{Path to the sub-catchment ID .tif layer}

\item{basin_path}{Path to the basin ID .tif layer}
}
\description{
Add the column 'subcID' to the dataset in 'dataset_path' and export the
data frame in a csv.
In the current version the long-lat values must be in the 2nd and 3rd
column of the data set text file.
}
