% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_intersection.R
\name{get_lake_intersection}
\alias{get_lake_intersection}
\title{Get intersection points between Hydrography90m stream network and lake shapefiles}
\usage{
get_lake_intersection(
  data,
  lake_id = "HydroLAKES_polys_v10",
  lakes,
  lake_name,
  buffer = TRUE,
  edge,
  stream,
  flow,
  basins,
  lake_dat,
  n_cores = 1,
  quiet = TRUE
)
}
\arguments{
\item{data}{a data.frame or data.table that contains the columns regarding
the lake ids, i.e. HydroLAKES ids. (i.e., output of extract_lake_ids);
see also help(extract_lake_ids)}

\item{lake_id}{character. The name of the column containing lake ids.}

\item{lakes}{character. Full path to lake geo-spatial files; i.e. HydroLAKES shapefiles.}

\item{lake_name}{character. The name of the attribute table column in
the geo-spatial files containing the lake ids; i.e. "HydroLAKES_polys_v10"
for HydroLAKES shapefiles.}

\item{buffer}{character. users can either set buffer to TRUE, FALSE or provide
buffer size in meters;
if TRUE a predefined buffer is used where the size of the buffer depends on
the lake size; if FALSE no buffer is applied; Default is TRUE}

\item{edge}{character. Full path to GuidosToolbox Workbench MSPA tool
see also (https://forest.jrc.ec.europa.eu/en/activities/lpa/gtb/)}

\item{stream}{character. Full path to Hydrography90m stream network tif file
add a reference were to check the necessary tif files for lake analysis}

\item{flow}{character. Full path to Hydrography90m flow accumulation tif file}

\item{basins}{character. Full path to Hydrography90m basin tif file}

\item{lake_dat}{character. Full path of the output.csv table,
i.e., the lake intersection table.}

\item{n_cores}{interger. Number of cores used in parallelization, Default is one.}

\item{quiet}{logical. If FALSE, the standard output will be printed.
Default is TRUE.}
}
\description{
The function identifies, for each lake ID, the intersection points
between the stream network (i.e., Hydrography90m) and the corresponding
lake's geospatial file (e.g., HydroLAKES shapefile).
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Note: For this we need the lake IDs (i.e. HydroLake ID), extracted from the
# extract_lake_ids.R function; see also help(extract_lake_ids)

data <- fread(paste0(my_directory,
                         "/hydrography90m_test_data",
                         "/lake_id.txt"),
                         header = TRUE)
lake_id <- "lake_id"

lakes <- (paste0(my_directory,
                    "/hydrography90m_test_data",
                    "/hydrography90m_test_lakes.gpkg"))

# To run the function we need to have installed GuidosToolbox Workbench MSPA tool
# see also (https://forest.jrc.ec.europa.eu/en/activities/lpa/gtb/) and
# (https://glowabio.github.io/hydrographr/articles/case_study_lake_workflow.html)
# give full path virtual machine linux path
edge <- (paste0(/home/USER/GWB_version/,
                    "/GWB"))

stream <- (paste0(my_directory,
                    "/hydrography90m_test_data",
                    "/stream_1264942.tif"))

flow <- (paste0(my_directory,
                    "/hydrography90m_test_data",
                    "/flow_1264942.tif"))

basins <- (paste0(my_directory,
                    "/hydrography90m_test_data",
                    "/basin_1264942.tif"))

lake_dat <- (paste0(my_directory,
                    "/hydrography90m_test_data"))

lake_intersect_table <- get_lake_intersection(data, lake_id = "lake_id",
                            lakes, lake_name = "lake_id", buffer = TRUE, edge,
                            stream, flow, basins, lake_dat, n_cores = 1,
                            quiet = FALSE)
}
\references{
https://grass.osgeo.org/grass82/manuals/
Soille P. and Vogt P. (2009). Morphological segmentation of binary patterns.
Pattern Recognition Letters 30, 4:456-459, doi: 1.1016/j.patrec.2008.10.015
\url{https://ies-ows.jrc.ec.europa.eu/gtb/GTB/MSPA_Guide.pdf}
}
\author{
Jaime Garcia Marquez, Thomas Tomiczek
}
