% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_intersection.R
\name{get_lake_intersection}
\alias{get_lake_intersection}
\title{Get intersection points between Hydrography90m stream network and lake shapefiles}
\usage{
get_lake_intersection(
  data,
  lake_id = "HydroLAKES_polys_v10",
  lakes,
  lake_name,
  buffer = TRUE,
  edge,
  stream,
  flow,
  glbasins,
  lake_dat,
  n_cores = 1,
  quiet = TRUE
)
}
\arguments{
\item{data}{a data.frame or data.table that contains the columns regarding
the lake ids, i.e. HydroLAKES ids.}

\item{lake_id}{character. The name of the column containing lake ids.}

\item{lakes}{character. Full path to lake geo-spatial files; i.e. HydroLAKES shapefiles.}

\item{lake_name}{character. The name of the attribute table column in
the geo-spatial files containing the lake ids; i.e. "HydroLAKES_polys_v10"
for HydroLAKES shapefiles.}

\item{buffer}{character. users can either set buffer to TRUE, FALSE or provide
buffer size in meters;
if TRUE a predefined buffer is used where the size of the buffer depends on
the lake size; if FALSE no buffer is applied; Default is TRUE}

\item{edge}{character. Full path to GuidosToolbox Workbench MSPA tool
see also (https://forest.jrc.ec.europa.eu/en/activities/lpa/gtb/)
(i.e., output of get_predict_table); see also help(get_predict_table)}

\item{stream}{character. Full path to Hydrography 90m stream network tif file
add a reference were to check the necessary tif files for lake analysis}

\item{flow}{character. Full path to Hydrography 90m flow accumulation tif file}

\item{glbasins}{character. Full path to Hydrography 90m basin tif file}

\item{lake_dat}{character. Full path of the output.csv table,
i.e., the lake intersection table.}

\item{n_cores}{interger. Number of cores used in parallelzation, Default is one.}

\item{quiet}{logical. If FALSE, the standard output will be printed.
Default is TRUE.}

\item{read}{logical. If TRUE, then the model .csv table
gets read into R as data.table and data.frame.
if FALSE, the table is only stored on disk. Default is FALSE.}
}
\description{
Add describtion text here

 Any thing the user needs to notice? add Note
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Get the lake intersection table for all lakes in the test data
# Note: For this we need the lake IDs (i.e. HydroLake ID), extracted from the
# extract_lake_ids.R function; see also help(extract_lake_ids)


}
\references{
add reference manual html here
https://grass.osgeo.org/grass82/manuals/
add manual of the MSPA analysis tool here
}
\author{
Jaime Garcia Marquez, Thomas Tomiczek
}
