% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pfafstetter_basins.R
\name{get_pfafstetter_basins}
\alias{get_pfafstetter_basins}
\title{Get Pfafstetter basins}
\usage{
get_pfafstetter_basins(
  g = g,
  subc_raster = subc_raster,
  out_dir = out_dir,
  file_name = file_name,
  data_table = FALSE,
  n_cores = n_cores
)
}
\arguments{
\item{g}{A directed graph of a basin with one outlet. The outlet can be any
stream / sub-catchment for which the upstream basin should be split
into smaller ones.}

\item{subc_raster}{Full path to the sub-catchment raster file of the basin.
Does not need to be cropped / masked to the basin.}

\item{out_dir}{The path of the output directory where the Pfafstetter
raster layer will be written.}

\item{file_name}{The filename and extension of the Pfafstetter raster
layer (e.g. 'pfafstetter_raster.tif")}

\item{data_table}{Logical. If TRUE, then the result will be loaded into R
as a 2-column data.table (sub-catchment ID and Pfafstetter code). If FALSE,
the result is provided as a raster (terra object) and written to disk.
Default is FALSE.}

\item{n_cores}{numeric. Number of cores used for parallelization. Default is
detectCores(logical=FALSE)-2}

\item{outlet}{The stream or sub-catchment ID that serves as the outlet of
the basin for which the Pfafstetter basins should be delineated.}
}
\value{
Either a data.table, or a raster (terra object) loaded into R. In
case the result is a raster, then a .tif file is written to disk.
}
\description{
Subset a basin or catchment into up to nine smaller units following the
Pfafstetter basin delineation scheme.
}
\details{
Each drainage basin can be split into smaller sub-basins following a
hierarchical topological coding scheme (see Verdin & Verdin (1999) for
details). This function splits a given basin into up to nine sub-basins
using the flow accumulation as the basis. The user has to define the
sub-catchment (stream segment) ID
}
\note{
You can use the online map at https://geo.igb-berlin.de/maps/351/view
to identify an ID of a stream segment (use the "Stream segment ID" layer to
the left)
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Import the stream network as a graph
g <- read_geopackage("order_vect_59.gpkg",
                      import_as ="graph")

# Subset the graph such that it contains only one one basin. You can use
# a random ID, i.e. it does not need to be the real outlet of the basin.
g <- get_catchment_graph(g,
                         stream=513867228,
                         outlet=FALSE,
                         mode="in",
                         as_graph = TRUE)

# Specify the sub-catchment raster file
subc_raster <- paste0(DIR, "/sub_catchment_h34v10.tif")

# Specify the output directory
out_dir <- my_directory

# Calculate the Pfafstetter sub-basins and write the raster layer to disk (
# and import into R)
pfafstetter <- get_pfafstetter_basins(g=my_graph,
                                      subc_raster=subc_raster,
                                      out_dir=out_dir,
                                      file_name="pfafstetter_raster.tif",
                                      data_table=FALSE,
                                      n_cores=4)


}
\references{
Verdin, K.L. & Verdin, J.P. (1999). A topological system
for delineation and codification of the Earth’s river basins. Journal
of Hydrology, 218(1-2), 1-12. doi:10.1016/s0022-1694(99)00011-6
}
\author{
Sami Domisch
}
