% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predict_table.R
\name{get_predict_table}
\alias{get_predict_table}
\title{Get predict table}
\usage{
get_predict_table(
  variable,
  statistics = "ALL",
  tile_id,
  input_var_path,
  subcatch_id,
  out_file_path,
  n_cores = NULL,
  read = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{variable}{character vector of variable names. Possible values are:}

\item{statistics}{character vector of statics names. Possible values are
"sd", "mean", "range" or "ALL". Default "ALL"}

\item{tile_id}{character. The IDs of the tiles of interest.}

\item{input_var_path}{path to table with environmental variables for entire
tiles.}

\item{subcatch_id}{path to a text file with subcatchments ids.}

\item{out_file_path}{character. The path to the output file.}

\item{n_cores}{numeric. Number of cores used for parallelization.}

\item{read}{logical. If TRUE, the table with environmental variables gets
read into R.
If FALSE, the table is only stored on disk. Default is TRUE.}

\item{quiet}{logical. If FALSE, the standard output will be printed.
Default is TRUE.}
}
\value{
The function returns...
}
\description{
This function creates a table with environmental variables from
an specific subset of subcatchments.
}
\examples{
bla
}
\author{
Jaime García Márquez, Yusdiel Torres-Cambas
s
}
