% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_unit_id.R
\name{get_regional_unit_id}
\alias{get_regional_unit_id}
\title{Adds sub-catchment and/or basin IDs to a dataframe of points}
\usage{
get_regional_unit_id(data, lon, lat)
}
\arguments{
\item{data}{data.frame with the lat lon columns}

\item{lon}{Name of the longitude column}

\item{lat}{Name of the latitude column}
}
\description{
Adds sub-catchment and/or basin IDs to a dataframe of points
}
