% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_unit_id.R
\name{get_regional_unit_id}
\alias{get_regional_unit_id}
\title{Get Hydrography90m regional unit IDs}
\usage{
get_regional_unit_id(data, lon, lat, quiet = TRUE)
}
\arguments{
\item{data}{a data.frame or data.table with lat/lon coordinates in WGS84.}

\item{lon}{character. The name of the column with the longitude coordinates.}

\item{lat}{character. The name of the column with the latitude coordinates.}

\item{quiet}{logical. If FALSE, the standard output will be printed.
Default is TRUE.}
}
\description{
Identifies the IDs of the regional units within the Hydrography90m data
in which the input points are located. The IDs are required to then download
the data using download_tiles().
Input is a point data frame.
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Read the species data
species_occurence <- read.table(paste0(my_directory,
                                       "/hydrography90m_test_data",
                                       "/spdata_1264942.txt"),
                              header = TRUE)

# Get the regional unit ID
get_regional_unit_id(species_occurence, lon = "longitude",
                    lat = "latitude")
}
\references{
\url{https://gdal.org/programs/gdallocationinfo.html}
}
\author{
Afroditi Grigoropoulou
}
