% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_location_info.R
\name{get_subcID}
\alias{get_subcID}
\title{Join sub-catchment IDs with occurrence points}
\usage{
get_subcID(taxon, subc_path, dataset_path)
}
\arguments{
\item{subc_path}{Description of parameter subc_path}

\item{dataset_path}{Description of parameter dataset_path}
}
\description{
Add the column 'subcID' to the dataset in 'dataset_path' and export the
dataframe in a csv.
The longitude-latitude values should be in the 3rd-4th column respectively!!
}
