% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_id.R
\name{get_tile_id}
\alias{get_tile_id}
\title{Get the Hydrography90m 20째x20째 tile ID}
\usage{
get_tile_id(data, lon, lat)
}
\arguments{
\item{data}{a data.frame or data.table that contains the columns regarding
the longitude / latitude coordinates in WGS84.}

\item{lon}{character. The name of the column with the longitude coordinates.}

\item{lat}{character. The name of the column with the latitude coordinates.}
}
\description{
Identifies the 20째x20째 tile IDs of the Hydrography90m
data in which the input points are located. The IDs can then be used to
download the data using \code{\link{download_tiles()}}. The input is a data
frame with point coordinates.
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Load species occurrence data
species_occurrence <- read.table(paste0(my_directory,
                                       "/hydrography90m_test_data",
                                       "/spdata_1264942.txt"),
                                 header = TRUE)

# Get the tile ID
get_tile_id(data = species_occurrence,
            lon = "longitude", lat = "latitude")
}
\author{
Afroditi Grigoropoulou
}
