% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_id.R
\name{get_tile_id}
\alias{get_tile_id}
\title{Identifies the ids of the tiles in which the given points are located.
Input is a point data frame.}
\usage{
get_tile_id(data, lon, lat)
}
\arguments{
\item{data}{data.frame with the lat lon columns}

\item{lon}{Name of the longitude column}

\item{lat}{Name of the latitude column}
}
\description{
Identifies the ids of the tiles in which the given points are located.
Input is a point data frame.
}
\author{
Afroditi Grigoropoulou
}
