% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upstream_catchment.R
\name{get_upstream_catchment}
\alias{get_upstream_catchment}
\title{Calculates the upstream basin taking each point as the outlet}
\usage{
get_upstream_catchment(
  data,
  id = character(0),
  lon = character(0),
  lat = character(0),
  direction_layer = NULL,
  out_path = NULL,
  n_cores = NULL,
  quiet = TRUE
)
}
\arguments{
\item{data}{Data.frame of snapped coordinates}

\item{id}{Name of the column containing the occurrence/site id}

\item{lon}{Name of the longitude column (WGS 84)}

\item{lat}{Name of the latitude column (WGS 84)}

\item{direction_layer}{Full path to raster file with the direction variable}

\item{out_path}{Full path to the directory where the output(s) will be stored}

\item{n_cores}{Numeric. Number of cores used for parallelization}

\item{quiet}{If FALSE, process is printed}
}
\description{
Calculates the upstream basin taking each point as the outlet
}
