% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tile.R
\name{merge_tiles}
\alias{merge_tiles}
\title{Merge raster or spatial vector tiles}
\usage{
merge_tiles()
}
\arguments{
\item{tile_path}{Path to the raster tile or spatial vector}
}
\description{
Merge raster or spatial vector tiles
}
