% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tile.R
\name{merge_tiles}
\alias{merge_tiles}
\title{Merge multiple raster or spatial vector objects}
\usage{
merge_tiles(tile_path, output_path, rraster_read = TRUE, rvector_read = FALSE)
}
\arguments{
\item{tile_path}{Path to directory containing the raster tiles or spatial vectors to be merged}

\item{output_path}{Path to write the output}

\item{rraster_read}{If TRUE merged raster .tif layer gets read into R.
TRUE is set by default}

\item{rvector_read}{If TRUE merged spatial vector gets read into R.
FALSE is set by default and if TRUE rraster_read needs to be set to FALSE}
}
\value{
A .tif raster file or spatial vector object that is always written
to disk, and optionally loaded into R.
}
\description{
Merge multiple raster or spatial vector objects from disk
to form a new raster or spatial vector object with a larger spatial extent.
A directory with at least a raster .tif or spatial vector geopackage file must be provided.
If rraster`_`read = TRUE (default), the outputs are a .tif
(saved under output`_`path) and a SpatRaster (terra package) object,
otherwise if rraster`_`read = FALSE, the .tif file is the only output.
If rvector`_`read = TRUE, the outputs are a .gpkg (saved under output`_`path)
and a SpatVector (terra package) object,
otherwise if rvector`_`read = FALSE (default),
the .tif file is the only output
}
