% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geopackage.R
\name{read_geopackage}
\alias{read_geopackage}
\title{Read a GeoPackage file}
\usage{
read_geopackage(
  file_name,
  type = NULL,
  SQL_table = NULL,
  as_dt = FALSE,
  as_graph = FALSE,
  as_sf = FALSE,
  as_SpatVect = FALSE
)
}
\arguments{
\item{file_name}{character. Name of the GeoPackage file to import,
e.g. "order_vect_segment_h00v00.gpkg".}

\item{type}{character. Either "net" for importing a network, "subc" for
sub-catchments, "basin" for drainage basins, or "outlet" for outlet points.}

\item{SQL_table}{character. Name of the specific data to import from the
GeoPackage. This is set automatically for the Hydrography90m data,
but needs to be specified for any other data. Optional.}

\item{as_dt}{logical. If TRUE, import the GeoPackage as a data.table.
Default is FALSE.}

\item{as_graph}{"logical. If TRUE, import the GeoPackage as a directed graph
(igraph object). Only possible for a network layer. Default is FALSE.}

\item{as_sf}{logical. If TRUE, import the GeoPackage as a spatial dataframe
(sf object). Default is FALSE.}

\item{as_SpatVect}{logical. If TRUE, import the GeoPackage as a
SpatVector (terra object). Default is FALSE.}
}
\description{
Reads a GeoPackage vector file from disk either as a table
(data.table), as a directed graph object (igraph), a spatial dataframe (sf)
or a SpatVect object (terra).
}
\examples{
library(hydrographr)

# Download test data into temporary R folder
my_directory <- tempdir()
download_test_data(my_directory)

# Read the stream network as a graph
my_graph <- read_geopackage(paste0(my_directory, "/order_vect_59.gpkg"),
type = "net", as_graph = T)

# Read the stream network as a data.table
my_dt <- read_geopackage(paste0(my_directory, "/order_vect_59.gpkg"),
type = "net", as_dt = T)

# Read the sub-catchments as a data.table
my_dt <- read_geopackage(paste0(my_directory, "/order_vect_59.gpkg"),
type = "subc", as_dt = T)

# Read the basin as a SF-object
my_sf <- read_geopackage(paste0(my_directory, "/order_vect_59.gpkg"),
type = "basin", as_sf = T)

#' # Read the basin as SpatVect object
my_sf <- read_geopackage(paste0(my_directory, "/order_vect_59.gpkg"),
type = "basin", as_SpatVect = T)
# Read the outlets as data.table
my_sf <- read_geopackage(paste0(my_directory, "/order_vect_59.gpkg"),
type = "outlet", as_sf = T)

}
\author{
Sami Domisch
}
