% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_no_data.R
\name{report_no_data}
\alias{report_no_data}
\title{Reports NoData value of raster layers.}
\usage{
report_no_data(data_dir, var_layer, n_cores = NULL)
}
\arguments{
\item{data_dir}{character. Path to the directory containing all input data.}

\item{var_layer}{character vector of variable raster layers on disk,
e.g. "slope_grad_dw_cel_h00v00.tif".}

\item{n_cores}{numeric. Number of cores used for parallelization, in case
multiple .tif files are provided to var_layer.}
}
\description{
This function reports the defined NoData value of a raster layer. The NoData
value of a raster layer represents the absence of data. In computations the
NoData value can be treated in different ways. Either the NoData value will
be reported or the Nodata value will be ignored and a value is computed from
the available values of a specified location.
}
\examples{
# Download test data into temporary R folder
my_directory <- tempdir()
download_test_data(my_directory)

# Report the NoData value
report_no_data(data_dir = paste0(my_directory, "/hydrography90m_test_data"),
               variable = c("subcatchment_1264942.tif", "flow_1264942.tif",
                             "spi_1264942.tif"),
               n_core = 2)

}
\references{
\url{https://gdal.org/programs/gdalinfo.html}
}
\author{
Afroditi Grigoropoulou, Maria M. Üblacker
}
