% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_no_data.R
\name{set_no_data}
\alias{set_no_data}
\title{Change or set no data value to raster layers.
The change happens in-place, meaning that the original file is overwritten}
\usage{
set_no_data(data_dir, variable, no_data)
}
\arguments{
\item{data_dir}{Character. Path to the directory containing all input data.}

\item{variable}{Character vector. Variable file name,
e.g. slope_grad_dw_cel_h00v00.tif. Variable names should remain
intact in file names, even after prior file processing,
i.e., slope_grad_dw_cel should appear in the file name
The files should be cropped to the extent of the sub-catchment layer}

\item{no_data}{The desired no data value}
}
\description{
Change or set no data value to raster layers.
The change happens in-place, meaning that the original file is overwritten
}
\author{
Afroditi Grigoropoulou
}
