% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_no_data.R
\name{set_no_data}
\alias{set_no_data}
\title{Set no data value}
\usage{
set_no_data(data_dir, var_layer, no_data, quiet = TRUE)
}
\arguments{
\item{data_dir}{character. Path to the directory containing all input data.}

\item{var_layer}{character vector of variable layers on disk,
e.g. c("sti_h16v02.tif", "slope_grad_dw_cel_h00v00.tif").
The original files will be overwritten.}

\item{no_data}{numeric. The desired NoData value.}

\item{quiet}{logical. If FALSE, the standard output will be printed.
Default is TRUE.}
}
\description{
Change or set the NoData value for a raster layer.
The change happens in-place, meaning that the original file is overwritten
on disk.
}
\examples{
# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Define no data value
set_no_data(data_dir = paste0(my_directory, "/hydrography90m_test_data"),
            var_layer = "cti_1264942.tif",
            no_data = -9999)

}
\references{
\url{https://gdal.org/programs/gdal_edit.html}
}
\author{
Afroditi Grigoropoulou, Maria M. Üblacker
}
