% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_to_network.R
\name{snap_to_network}
\alias{snap_to_network}
\title{Snap data points to the next stream segment within a defined radius
or a minimum flow accumulation.}
\usage{
snap_to_network(
  data,
  lon,
  lat,
  site_id,
  stream_path,
  accu_path = NULL,
  calc = "dist",
  dist = 500,
  accu = 0.5,
  quiet = TRUE
)
}
\arguments{
\item{data}{Data.frame with lat/lon coordinates in WGS84.}

\item{lon}{Column name of longitude coordinates as character vector.}

\item{lat}{Column name of latitude coordinates as character vector.}

\item{site_id}{Column name of a unique ID as character vector; Optional
an ID for the data points can be defined.}

\item{stream_path}{Full path of the stream network .tif file}

\item{accu_path}{Full path of the flow accumulation .tif file. Needed if
the point should be snapped to next stream segment with a value higher than
the flow accumulation threshold. This avoids points to snap to small stream
tributaries.}

\item{calc}{Options "dist", "accu", or "both". Default is "dist".
Defines if the points are snapped using the distance or flow accumulation.
If calc is set to "both" the output will contain the new coordinates for both
calculations.}

\item{dist}{Maximum radius in meters; Point will be snapped to the next
stream within a defined radius. By default the distance is 500 meter.}

\item{accu}{Minimum flow accumulation; Point will be snapped to the next
stream with a flow accumulation equal or higher than the given value. By
default the flow accumulation value is set to 0.5.}

\item{quiet}{Logical; Whether the standard output will be printed or not.}
}
\description{
Snap data points to the next stream segment within a defined radius
or a minimum flow accumulation.
}
\details{
The function makes use of the r.stream.snap command available in GRASS GIS to
simultaneously a number of points to a stream network. A distance threshold
can be specified and points will be snap to any stream segemnt within this
distance radius. However, to avoid snapping to small tributaries, an
acumulation threshold can be used and the snapping occurs on stream segment
with equal or higher accumulation threshold and within the given distance radius.
}
\note{
Duplicated rows will be removed.
}
\examples{
# Download test data into temporary R folder
my_directory <- tempdir()
download_test_data(my_directory)

# Load occurrence data
species_occurence <- read.table(paste0(my_directory,
                                "/hydrography90m_test_data/spdata_1264942.txt"),
                                header = TRUE

# Define full path to flow accumulation
stream_rast <- paste0(my_directory,
                     "/hydrography90m_test_data/stream_1264942.tif")
flow_rast <- paste0(my_directory,
                     "/hydrography90m_test_data/flow_1264942.tif")

# To calculate the new (snapped) coordinates for a radius and a flow
snapped_coordinates <- snap_to_network(data = species_occurrence,
                                       lon = "longitude",
                                       lat = "latitude",
                                       site_id = "occurrence_id",
                                       stream_path = stream_vect,
                                       accu_path = flow_rast,
                                       calc = "both",
                                       dist = 300,
                                       accu = 0.8)

# Show head of output table
head(snapped_coordinates)

}
\references{
\link[https://grass.osgeo.org/grass78/manuals/addons/r.stream.snap.html]
}
\author{
Maria M. Üblacker, Jaime Garcia Marquez
}
